==> processing/01-prod_prep-individual.Rmd <==
---
title: "Data preparation - individual"
date: "`r format(Sys.time(), '%A %d %B %Y %X')`"
output: 
  html_document: 
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: no
      number_sections: yes
    code_folding: show  
    number_sections: yes
editor_options: 
  chunk_output_type: console
---

# Setup

```{r setup}
options(scipen=9999) 
```

```{r}
if (!require("pacman")) install.packages("pacman") # instalar pacman
pacman::p_load(dplyr,sjmisc,GGally,sjlabelled,tidyverse,MLMusingR)
df1 <- haven::read_dta(here::here("input/data/original/ZA6980_v2-0-0.dta"))
dfTR <- haven::read_dta(here::here("input/data/original/ZA5521_v1-0-0.dta"))
dfES <- haven::read_dta(here::here("input/data/original/ZA5962_v1-0-0.dta"))
df1 <- df1 %>% rename(topbot=TOPBOT)
dfES <- dfES %>% rename(topbot=TOPBOT)
# fix problem with variable names
dfTR <- dfTR %>% rename(SEX=sex,
                        DEGREE=degree,
                        EDUCYRS=educyrs,
                        AGE=age,
                        MAINSTAT=mainstat,
                        UNION=union,
                        WEIGHT=weight,
                        URBRURAL=urbrural,
                        WORK=work,
                        HOMPOP=hompop,
                        EMPREL=emprel,
                        NSUP=nsup,
                        WRKSUP=wrksup,
                        SPEMPREL=spemprel,
                        SPWRKSUP=spwrksup,
                        ATTEND=attend, 
                        RELIGGRP=religgrp,
                        MARITAL=marital,
                        WRKHRS=wrkhrs)

dfES$c_alphan <- NULL
dfTR$c_alphan <- NULL
df2 <- bind_rows(dfES,dfTR)
df1<- bind_rows(df1,df2)
```

# Inequality perceptions

```{r}
df1$redist <- car::recode(df1$v13,recodes = "c(0,8,9)=NA")
df1$redist <- set_na(df1$redist,na = c(0,8,9))
df1$redist <- sjmisc::rec(x = df1$redist,rec = "rev")
sjlabelled::set_label(df1$redist) <- "Preferences goverment redistribution"

df1$prefineq <- car::recode(df1$v12,recodes = "c(0,8,9)=NA")
df1$prefineq <- set_na(df1$prefineq,na = c(0,8,9))
df1$prefineq <- sjmisc::rec(x = df1$prefineq,rec = "rev")
sjlabelled::set_label(df1$prefineq) <- "Preference inequality (society has to be fair)"

df1$pertransfer <- car::recode(df1$v14,recodes = "c(0,8,9)=NA")
df1$pertransfer <- set_na(df1$pertransfer,na = c(0,8,9))
df1$pertransfer <- sjmisc::rec(x = df1$pertransfer,rec = "rev")
sjlabelled::set_label(df1$pertransfer) <- "Perception about welfare (people become lazy)"
```

# Index egalitarian attitudes
```{r}
df1$egal2<- rowMeans(df1[,c("redist","prefineq")],na.rm = T)
df1$egal2 <- ((df1$egal2-min(df1$egal2,na.rm = T))/(max(df1$egal2,na.rm = T)-min(df1$egal2,na.rm = T)))*100
sjlabelled::set_label(df1$egal2) <- "Redistributive attitudes (2 items)"
```

# Social status

```{r income-hhold}
income_country <- names(df1 %>% dplyr::select(dplyr::ends_with("_INC")))
# lapply(df1[,income_country], table)
for (i in income_country) {
  # df1[[i]] <- car::recode(var = df1[[i]],"999990:99999999=NA")
  df1[[i]] <- as.numeric(df1[[i]])
  }

df1$AT_INC<- car::recode( df1$AT_INC,"999990:999999=NA")
df1$AU_INC<- car::recode( df1$AU_INC,"9999990:9999999=NA")
df1$CH_INC<- car::recode( df1$CH_INC,"999990:999999=NA")
df1$CN_INC<- car::recode( df1$CN_INC,"9999990:9999999=NA")
df1$CZ_INC<- car::recode( df1$CZ_INC,"999990:999999=NA")
df1$DE_INC<- car::recode( df1$DE_INC,"999990:999999=NA")
df1$DK_INC<- car::recode( df1$DK_INC,"9999990:9999999=NA")
df1$ES_INC<- car::recode( df1$ES_INC,"999990:999999=NA")
df1$FI_INC<- car::recode( df1$FI_INC,"999990:999999=NA")
df1$FR_INC<- car::recode( df1$FR_INC,"999990:999999=NA")
df1$GB_INC<- car::recode( df1$GB_INC,"999990:999999=NA")
df1$HR_INC<- car::recode( df1$HR_INC,"999990:999999=NA")
df1$HU_INC<- car::recode( df1$HU_INC,"9999990:9999999=NA")
df1$IL_INC<- car::recode( df1$IL_INC,"999990:999999=NA")
df1$IN_INC<- car::recode( df1$IN_INC,"999990:999999=NA")
df1$IS_INC<- car::recode( df1$IS_INC,"99999990:99999999=NA")
df1$JP_INC<- car::recode( df1$JP_INC,"99999990:99999999=NA")
df1$LT_INC<- car::recode( df1$LT_INC,"999990:999999=NA")
df1$MX_INC<- car::recode( df1$MX_INC,"999990:999999=NA")
df1$NZ_INC<- car::recode( df1$NZ_INC,"999990:999999=NA")
df1$PH_INC<- car::recode( df1$PH_INC,"999990:999999=NA")
df1$RU_INC<- car::recode( df1$RU_INC,"999990:999999=NA")
df1$SE_INC<- car::recode( df1$SE_INC,"999990:999999=NA")
df1$SI_INC<- car::recode( df1$SI_INC,"999990:999999=NA")
df1$SK_INC<- car::recode( df1$SK_INC,"999990:999999=NA")
df1$SR_INC<- car::recode( df1$SR_INC,"999990:999999=NA")
df1$TH_INC<- car::recode( df1$TH_INC,"999990:999999=NA")
df1$TW_INC<- car::recode( df1$TW_INC,"9999990:9999999=NA")
df1$US_INC<- car::recode( df1$US_INC,"999990:999999=NA")
df1$ZA_INC<- car::recode( df1$ZA_INC,"999990:999999=NA")

frq(df1$EE_INC)

c((200 )/1,
    (201 + 300)/2,
    (301 + 400)/2,
    (401 + 500)/2,
    (501 + 650)/2,
    (651 + 800)/2,
    (801 + 1000)/2,
    (1001 +1300)/2,
    (1301 + 1600)/2,
    (1601 +2100)/2,
    (2500)/1)

df1$EE_INC<- car::recode( df1$EE_INC,"1=200.0;2=250.5;3=350.5;4=450.5;5=575.5;6=725.5;7=900.5;8=1150.5;9=1450.5;10=1850.5;11=2500.0;999990:999999=NA")

frq(df1$TR_INC)
df1$TR_INC<- car::recode( df1$TR_INC,"999990:999997=NA")
# lapply(df1[,income_country], table)  # Ok el recode de ingresos familiares 
# Crear variable ingresos unica 
df1 <- df1 %>% 
  mutate(income = coalesce(AT_INC,AU_INC,CH_INC,CN_INC,CZ_INC,DE_INC,DK_INC,ES_INC,
                           FI_INC,FR_INC,GB_INC,HR_INC,HU_INC,IL_INC,IN_INC,IS_INC,
                           JP_INC,LT_INC,MX_INC,NZ_INC,PH_INC,RU_INC,SE_INC,SI_INC,
                           SK_INC,SR_INC,TH_INC,TW_INC,US_INC,ZA_INC,EE_INC,TR_INC))



table(df1$c_alphan,is.na(df1$income))

# summary(df1$JP_RINC)
# summary(df1$JP_RINC)
# 
# summary(df1$EE_INC)
# summary(df1$TR_INC)

#frq(df1$income==0)
frq(df1$HOMPOP)
df1$hhpop <- as.numeric(df1$HOMPOP) 
df1$hhpop[df1$hhpop %in%c(0,97,99)] <- NA
# frq(df1$hhpop)
df1$incomepcap <- as.numeric(df1$income/df1$hhpop) # Ingreso per capita del hogar
  
df1 <- df1 %>%group_by(country)  %>% mutate(D10pc=ntile(incomepcap,10),
                                              Q05pc=ntile(incomepcap,5),
                                              Q03pc=ntile(incomepcap,3),
                                              zincpc=scale(incomepcap,center = TRUE),
                                              logincpc=log(incomepcap)) %>% ungroup()

df1$Q05pcm <- car::recode(df1$Q05pc,"1='Q01';2='Q02';3='Q03';4='Q04';5='Q05';NA='Missing'",
                          as.factor = T,
                          levels = c("Q01","Q02","Q03","Q04","Q05","Missing")) 
frq(df1$Q05pc)

df1$Q03pcm <- car::recode(df1$Q03pc,"1='T01';2='T02';3='T03';NA='Missing'",
                          as.factor = T,
                          levels = c("T01","T02","T03","Missing")) 
frq(df1$Q03pcm)

table(as_character(df1$country),df1$Q03pcm)

sjlabelled::set_label(df1$D10pc) <- "Household income (D10)"
sjlabelled::set_label(df1$Q05pc) <- "Household income (Q05)"
sjlabelled::set_label(df1$zincpc) <- "Household income (zscore)"
sjlabelled::set_label(df1$logincpc) <- "Household income (log)"
```

```{r income-respondent}
income_country <- names(df1 %>% dplyr::select(dplyr::ends_with("_RINC")))
# lapply(df1[,income_country], table)
for (i in income_country) {
  # df1[[i]] <- car::recode(var = df1[[i]],"999990:99999999=NA")
  df1[[i]] <- as.numeric(df1[[i]])
  }


df1$AT_RINC<- car::recode( df1$AT_RINC,"999990:999999=NA")
df1$AU_RINC<- car::recode( df1$AU_RINC,"999990:999999=NA")
df1$CH_RINC<- car::recode( df1$CH_RINC,"999990:999999=NA")
df1$CN_RINC<- car::recode( df1$CN_RINC,"9999990:9999999=NA")
df1$CZ_RINC<- car::recode( df1$CZ_RINC,"999990:999999=NA")
df1$DE_RINC<- car::recode( df1$DE_RINC,"999990:999999=NA")
df1$DK_RINC<- car::recode( df1$DK_RINC,"999990:999999=NA")
df1$ES_RINC<- car::recode( df1$ES_RINC,"999990:999999=NA")
df1$FI_RINC<- car::recode( df1$FI_RINC,"999990:999999=NA")
df1$FR_RINC<- car::recode( df1$FR_RINC,"999990:999999=NA")
df1$GB_RINC<- car::recode( df1$GB_RINC,"999990:999999=NA")
df1$HR_RINC<- car::recode( df1$HR_RINC,"999990:999999=NA")
df1$HU_RINC<- car::recode( df1$HU_RINC,"999990:999999=NA")
df1$IL_RINC<- car::recode( df1$IL_RINC,"999990:999999=NA")
df1$IN_RINC<- car::recode( df1$IN_RINC,"999990:999999=NA")
df1$IS_RINC<- car::recode( df1$IS_RINC,"9999990:9999999=NA")
df1$JP_RINC<- car::recode( df1$JP_RINC,"99999990:99999999=NA")
df1$LT_RINC<- car::recode( df1$LT_RINC,"999990:999999=NA")
df1$MX_RINC<- car::recode( df1$MX_RINC,"999990:999999=NA")
df1$NZ_RINC<- car::recode( df1$NZ_RINC,"999990:999999=NA")
df1$PH_RINC<- car::recode( df1$PH_RINC,"999990:999999=NA")
df1$RU_RINC<- car::recode( df1$RU_RINC,"999990:999999=NA")
df1$SE_RINC<- car::recode( df1$SE_RINC,"999990:999999=NA")
df1$SI_RINC<- car::recode( df1$SI_RINC,"999990:999999=NA")
df1$SK_RINC<- car::recode( df1$SK_RINC,"999990:999999=NA")
df1$SR_RINC<- car::recode( df1$SR_RINC,"999990:999999=NA")
df1$TH_RINC<- car::recode( df1$TH_RINC,"999990:999999=NA")
df1$TW_RINC<- car::recode( df1$TW_RINC,"999990:999999=NA")
df1$US_RINC<- car::recode( df1$US_RINC,"999990:999999=NA")
df1$ZA_RINC<- car::recode( df1$ZA_RINC,"999990:999999=NA")

frq(df1$EE_RINC)
df1$EE_RINC<- car::recode( df1$EE_RINC,"1=200.0;2=250.5;3=350.5;4=450.5;5=575.5;6=725.5;7=900.5;8=1150.5;9=1450.5;10=1850.5;11=2500.0;999990:999999=NA")
frq(df1$EE_RINC)

frq(df1$TR_RINC)
df1$TR_RINC<- car::recode( df1$TR_RINC,"999990:999997=NA")
frq(df1$TR_RINC)


# lapply(df1[,income_country], table)  # Ok el recode de ingresos familiares 
# Crear variable ingresos unica 
df1 <- df1 %>% mutate(incomer = coalesce(AT_RINC,AU_RINC,
                                        CH_RINC, CN_RINC,
                                        CZ_RINC, DE_RINC,
                                        DK_RINC, ES_RINC,
                                        FI_RINC, FR_RINC,
                                        GB_RINC, HR_RINC,
                                        HU_RINC, IL_RINC,
                                        IN_RINC, IS_RINC,
                                        JP_RINC, LT_RINC,
                                        MX_RINC, NZ_RINC,
                                        PH_RINC, RU_RINC,
                                        SE_RINC, SI_RINC,
                                        SK_RINC, SR_RINC,
                                        TH_RINC, TW_RINC,
                                        US_RINC, ZA_RINC, 
                                        EE_RINC,TR_RINC))
table(df1$c_alphan,is.na(df1$income))
prop.table(table(df1$c_alphan,is.na(df1$income)),margin =1)


df1 <- df1 %>%group_by(country) %>% mutate(D10r=ntile(incomer,10),
                                              Q05r=ntile(incomer,5),
                                           Q03r=ntile(incomer,3),
                                              zincpcr=scale(incomer,center = TRUE),
                                              logincr=log(incomer)) %>% ungroup()


df1$Q05rm <- car::recode(df1$Q05r,"1='Q01';2='Q02';3='Q03';4='Q04';5='Q05';NA='Missing'",
                          as.factor = T,
                          levels = c("Q01","Q02","Q03","Q04","Q05","Missing")) 

df1$Q03rm <- car::recode(df1$Q03r,"1='T01';2='T02';3='T03';NA='Missing'",
                          as.factor = T,
                          levels = c("T01","T02","T03","Missing")) 

table(as_character(df1$country),df1$Q03rm)

# frq(df1$incomer)
# frq(df1$incomepcap)
# Single household 
df1$incomerep <- ifelse(df1$hhpop==1 & is.na(df1$incomepcap),df1$incomer,df1$incomepcap)
#frq(is.na(df1$incomerep))
#frq(df1$incomerep==0)
df1[,c("incomer","hhpop","incomepcap","incomerep")] %>% arrange(incomepcap,hhpop) %>% head()
df1$incomerep <- ifelse(df1$hhpop==1 & df1$incomerep==0,df1$incomer,df1$incomerep)
#frq(df1$incomerep==0)
df1[,c("incomer","hhpop","incomepcap","incomerep")] %>% arrange(incomepcap,hhpop) %>% head()

summary(df1[,c("incomer","hhpop","incomepcap","incomerep")])
```

```{r education}
#frq(df1$EDUCYRS)
df1$edyears <- as.numeric(car::recode(df1$EDUCYRS,recodes = "98:99=NA")) 
sjlabelled::set_label(df1$edyear) <- "Years of schooling"
#frq(df1$edyears)

#frq(df1$DEGREE)
df1$ednum <- car::recode(df1$DEGREE+1,"10=NA")
sjlabelled::set_label(df1$ednum) <- "Educational level"
#frq(df1$ednum)
df1$edcat <- car::recode(df1$ednum,recodes = "1:2='Primary or less';
                         3:4='Secondary';
                         5='Terciary short-cycle';
                         6:7='Terciary or higher';
                         9=NA",
                         levels =c('Primary or less',
                                   'Secondary',
                                   'Post-secondary non-terciary',
                                   'Terciary or higher'),as.factor = T)
sjlabelled::set_label(df1$edcat) <- "Educational level"
#frq(df1$edcat)
```

# Occupational variables

```{r labour-status}
#frq(df1$MAINSTAT)
df1$labst <- car::recode(
  as.numeric(df1$MAINSTAT),
  recodes = "1='In labour force';2:9='Not in labour force';c(0,99)='No information'",
  levels = c('In labour force', 'Not in labour force', 'No information'),
  as.factor = T
)
sjlabelled::set_label(df1$labst) <- "Occupational status"
#frq(df1$labst)
```

```{r paid-work}
frq(df1$WORK)
df1$workst <- car::recode(
  as.numeric(df1$WORK),
  recodes = "1='In paid work';2:3='Not in paid work';c(9)=NA",
  levels = c('In paid work', 'Not in paid work'),
  as.factor = T
)
sjlabelled::set_label(df1$workst) <- "Work status"
frq(df1$workst)
```

```{r isei}
df1$ISCO08[df1$ISCO08 %in% c(9998,9999)] <- NA
df1$isei08 <- occupar::isco08toISEI08(isco08 = df1$ISCO08,display.nas = T)
# summary(df1$isei08)
sjlabelled::set_label(df1$isei08) <- "ISEI respondent"


#frq(df1$SPISCO08)
df1$SPISCO08[df1$SPISCO08 %in% c(9998,9999)] <- NA
df1$isei08sp <- occupar::isco08toISEI08(isco08 = df1$SPISCO08,display.nas = T)
# summary(df1$isei08sp)
sjlabelled::set_label(df1$isei08sp) <- "ISEI spouse"
```

```{r class-variables}
## Ocupación
# Ocupación u oficio actual ISCO 08
#frq(df1$SPISCO08) # isco08 
# Recode CIUO_encuestado desde ISCO 2008 a ISCO 1988
df1$isco88spo <- occupar::isco08to88(df1$SPISCO08)
#frq(df1$isco88spo)

#Relación de Empleo encuestado
#frq(df1$SPEMPREL)

df1$sempspo <- df1$SPEMPREL
df1$sempspo <- as.numeric(car::recode(df1$sempspo, "c(2,3,4,5)=1;c(1)=2;c(0,8,9)=NA"))
#frq(df1$sempspo)

df1$sempspo <- sjlabelled::set_labels(x = df1$sempspo,labels=c('Self-employed','Employee'))
#frq(df1$sempspo)

# superivison
#frq(df1$SPWRKSUP) #supervision spouse
df1$supvisspo <- as.numeric(car::recode(df1$SPWRKSUP,"c(0,2)=0;1=9;c(8,9)=NA"))
#frq(df1$supvisspo)

## Ocupación
# Ocupación u oficio actual ISCO 08
#frq(df1$ISCO08) # isco08 
# Recode CIUO_encuestado desde ISCO 2008 a ISCO 1988
df1$isco88r <- occupar::isco08to88(df1$ISCO08)
#frq(df1$isco88r)

#Note: the package generate identical recodes as my own manual recoding

#Employment relation encuestado
frq(df1$EMPREL)
# frq(df1$WORK)
df1$semp <- df1$EMPREL
df1$semp <- as.numeric(car::recode(df1$semp, "c(2,3,4)=1;c(1)=2;c(0,8,9)=NA"))
frq(df1$semp)
df1$semp <- sjlabelled::set_labels(x = df1$semp,labels=c('Self-employed','Employee'))
#frq(df1$semp)

# superivison
frq(df1$NEMPLOY) #supervision respondent
frq(df1$NSUP)
df1$supvis <- as.numeric(car::recode(df1$NEMPLOY,"c(9998,9999)=NA"))

frq(df1$supvis)

# save the occupational variables for using them in stata-----------------------
df1$id <- NULL
df1 <- tibble::rowid_to_column(df1, "id")

# For comparing with the DIGCLASS

frq(df1$WRKSUP)
frq(df1$EMPREL)
frq(df1$NSUP)

df1 <-
  df1 %>%
  mutate(
    is_supervisor = ifelse(WRKSUP == 2, 0, as.numeric(car::recode(df1$WRKSUP,"c(8,9)=NA"))),
    self_employed = case_when(
      EMPREL %in% c(1,4) ~ 0,
      EMPREL %in% 2:3 ~ 1
    ),
    n_employees = ifelse(is_supervisor == 0, 0, as.numeric(car::recode(df1$NSUP,"c(9998,9999)=NA")))
  )

frq(df1$is_supervisor)
frq(df1$self_employed)
frq(df1$n_employees)


frq(df1$SPWORK)
frq(df1$SPEMPREL)

df1$NSUPSPO <- 1
frq(df1$NSUPSPO)

df1 <-
  df1 %>%
  mutate(
    is_supervisor_spo = ifelse(SPWORK == 2, 0, as.numeric(car::recode(df1$SPWORK,"c(8,9)=NA"))),
    self_employed_spo = case_when(
      SPEMPREL %in% c(1,4) ~ 0,
      SPEMPREL %in% 2:3 ~ 1
    ),
    n_employees_spo = ifelse(is_supervisor_spo == 0, 0, as.numeric(car::recode(df1$NSUPSPO,"c(9998,9999)=NA")))
  )

frq(df1$is_supervisor_spo)
frq(df1$self_employed_spo)
frq(df1$n_employees_spo)
```

# Social Class

```{r digclass-egp}
# install.packages("devtools") 
# devtools::install_git("https://code.europa.eu/digclass/digclass.git")
library("DIGCLASS")
df1 <- 
df1 %>% 
mutate(egp11 = isco88_to_egp(isco88r, self_employed, n_employees, label = F,to_factor = T),
       egp11spo = isco88_to_egp(isco88spo, self_employed_spo, n_employees_spo, label = F,to_factor = T))

frq(df1$egp11)
df1$egp11 <- as_numeric(df1$egp11)
frq(df1$egp11)

frq(df1$egp11spo)
df1$egp11spo <- as_numeric(df1$egp11spo)
frq(df1$egp11spo)

#if NA in class of respondent, replace with spouse
df1$dclass11 <- ifelse(is.na(df1$egp11),df1$egp11spo,df1$egp11)

frq(df1$dclass11)
#0.1 label respondent egp11 classes
df1$egp11 <- 
  sjlabelled::set_labels(df1$egp11,
                         levels = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10,11),
                         labels =  c("I", "II", "IIIa","IIIb", "IVa", "IVb", 
                                     "IVc", "V", "VI", "VIIa", "VIIb")) 

#0.2 label spouse egp11 classes
df1$egp11spo <- 
  sjlabelled::set_labels(df1$egp11spo,
                         levels = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10,11),
                         labels =  c("I", "II", "IIIa","IIIb", "IVa", "IVb", 
                                     "IVc", "V", "VI", "VIIa", "VIIb")) 

# if NA in respondents it is assumed to be lower than spouse
# now, with the cases in we have information for respondent and spouse
# we replace the value by the highest one, just for those

# Replace the highest class by dominance criteria 
# Keep the highest of two values:
df1$dclass11d <- pmax(df1$egp11, #social class respondent + spouse vs:
                     df1$egp11spo,#social class spouse
                     na.rm = T) # if NA, do nothing
frq(df1$dclass11d)
df1$dclass11d <- as_numeric(df1$dclass11d)
frq(df1$dclass11d)

df1$dclass11d <- 
  sjlabelled::set_labels(df1$dclass11d,
                         levels = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10,11),
                         labels =  c("I", "II", "IIIa","IIIb", "IVa", "IVb", 
                                     "IVc", "V", "VI", "VIIa", "VIIb")) 

sjlabelled::set_label(df1$dclass11d) <- "Social Class 11 Dominance (EGP11 -  DIGICLASS)";frq(df1$dclass11d)

# Class scheme using dominance-------------------------------------------------
# # Class Scheme 3 classes -  V to intermediate class
df1$dclass3 <- car::recode(df1$dclass11d,as.factor = T,
                            levels =c('Service Class (I+II)',
                                      'Intermediate class (III+IV+V)',
                                      'Working Class (VI+VII)'), 
                            "1:2='Service Class (I+II)';
                             3:8='Intermediate class (III+IV+V)';
                             9:11='Working Class (VI+VII)'")

sjlabelled::set_label(df1$dclass3) <- "Social Class 3 Dominance (V to Intermediate Class-  DIGICLASS)";frq(df1$dclass3)

# Class Scheme 3 classes -  V to working class
df1$dclass3_V <- car::recode(df1$dclass11d,as.factor = T,
                            levels =c('Service Class (I+II)',
                                      'Intermediate class (III+IV)',
                                      'Working Class (V+VI+VII)'), 
                            "1:2='Service Class (I+II)';
                             3:7='Intermediate class (III+IV)';
                             8:11='Working Class (V+VI+VII)'")

sjlabelled::set_label(df1$dclass3_V) <- "Social Class 3 Dominance (V to Working Class - DIGCLASS)";frq(df1$dclass3_V)


# Class Scheme 3 classes -  IIIb to working class
df1$dclass3_III <- car::recode(df1$dclass11d,as.factor = T,
                            levels =c('Service Class',
                                      'Intermediate class',
                                      'Working Class'), 
                            "c(1,2)='Service Class';c(3,5,6,7,8)='Intermediate class';c(4,9,10,11)='Working Class'")

sjlabelled::set_label(df1$dclass3_III) <- "Social Class 3 Dominance (IIIb to Working Class - DIGCLASS)";frq(df1$dclass3_III)

# Class Scheme 3 classes -  IIIb and V to working class
df1$dclass3_III_V <- car::recode(df1$dclass11d,as.factor = T,
                            levels =c('Service Class',
                                      'Intermediate class',
                                      'Working Class'), 
                            "c(1,2)='Service Class';c(3,5,6,7)='Intermediate class';c(4,8,9,10,11)='Working Class'")

sjlabelled::set_label(df1$dclass3_III_V) <- "Social Class 3 Dominance (IIIb+V to Working Class - DIGCLASS)";frq(df1$dclass3_III)

# Class Scheme 6 classes
df1$dclass6 <- car::recode(
  df1$dclass11d,
  as.factor = T,
  levels = c(
    'Service class (I - higher grade)',
    'Service class (II - lower grade)',
    'Routine non-manual (IIIa+b)',
    'Petty-bourgeoise (IVa+b+c)',
    'Skilled manual workers and supv. (V+VI)',
    'Non-skilled manual workers (VIIa+b)'
  ),
  "1='Service class (I - higher grade)';
   2='Service class (II - lower grade)';
   3:4='Routine non-manual (IIIa+b)';
   5:7='Petty-bourgeoise (IVa+b+c)';
   8:9='Skilled manual workers and supv. (V+VI)';
   10:11='Non-skilled manual workers (VIIa+b)'"
)

sjlabelled::set_label(df1$dclass6) <- "Social Class 6 Dominance (DIGCLASS)"
```



```{r digclass-egp-respondent}
# Class scheme using Respondent-------------------------------------------------
# # Class Scheme 3 classes -  V to intermediate class

df1$dclass3res <- car::recode(df1$egp11,as.factor = T,
                           levels =c('Service Class (I+II)',
                                     'Intermediate class (III+IV+V)',
                                     'Working Class (VI+VII)'), 
                           "1:2='Service Class (I+II)';
                             3:8='Intermediate class (III+IV+V)';
                             9:11='Working Class (VI+VII)'")

sjlabelled::set_label(df1$dclass3res) <- "Social Class 3 (R's) - (V to Intermediate Class-  DIGICLASS)";frq(df1$dclass3res)

# Class Scheme 3 classes -  V to working class
df1$dclass3res_V <- car::recode(df1$egp11,as.factor = T,
                             levels =c('Service Class (I+II)',
                                       'Intermediate class (III+IV)',
                                       'Working Class (V+VI+VII)'), 
                             "1:2='Service Class (I+II)';
                             3:7='Intermediate class (III+IV)';
                             8:11='Working Class (V+VI+VII)'")

sjlabelled::set_label(df1$dclass3res_V) <- "Social Class 3 (R's) - (V to Working Class - DIGCLASS)";frq(df1$dclass3res_V)


# Class Scheme 3 classes -  IIIb to working class
df1$dclass3res_III <- car::recode(df1$egp11,as.factor = T,
                               levels =c('Service Class',
                                         'Intermediate class',
                                         'Working Class'), 
                               "c(1,2)='Service Class';c(3,5,6,7,8)='Intermediate class';c(4,9,10,11)='Working Class'")

sjlabelled::set_label(df1$dclass3res_III) <- "Social Class 3 (R's) - (IIIb to Working Class - DIGCLASS)";frq(df1$dclass3res_III)

# Class Scheme 3 classes -  IIIb and V to working class
df1$dclass3res_III_V <- car::recode(df1$egp11,as.factor = T,
                                 levels =c('Service Class',
                                           'Intermediate class',
                                           'Working Class'), 
                                 "c(1,2)='Service Class';c(3,5,6,7)='Intermediate class';c(4,8,9,10,11)='Working Class'")

sjlabelled::set_label(df1$dclass3res_III_V) <- "Social Class 3 (R's) - (IIIb+V to Working Class - DIGCLASS)";frq(df1$dclass3res_III_V)

# Class scheme 6 classes 
df1$dclass6res <- car::recode(df1$egp11,as.factor = T,
                                 levels =c('Upper Service class',
                                           'Lower Service class',
                                           'Routine nonmanual class',
                                           'Self-employed',
                                           'Skilled working class',
                                           'Unskilled working class'), 
                                 "1='Upper Service class';2='Lower Service class';c(3,4)='Routine nonmanual class';c(5,6,7)='Self-employed';
                              c(8,9)='Skilled working class';c(10,11)='Unskilled working class'")

sjlabelled::set_label(df1$dclass6res) <- "Social Class 6 (R's) - (DIGCLASS)";frq(df1$dclass6res)
```

# Social Networks

```{r generator}
ocupations <- names(dplyr::select(df1,v1:v10))
for (i in ocupations) {
  df1[[i]] <- car::recode(df1[[i]],"c(8,9)=NA")
  df1[[i]] <- sjlabelled::set_na(df1[[i]],na = c(8,9))
}

# frq(df1$v1)
df1$busdriver <- as.numeric(car::recode(df1$v1,"1:3=1;4=0")) 
sjlabelled::set_label(df1$busdriver) <- "Know: bus/lorry driver" 
# frq(df1$busdriver)

#frq(df1$v2)
df1$ceo <- as.numeric(car::recode(df1$v2,"1:3=1;4=0")) 
sjlabelled::set_label(df1$ceo) <- "Know: CEO large company" 
#frq(df1$ceo)

#frq(df1$v3)
df1$cleaner <- as.numeric(car::recode(df1$v3,"1:3=1;4=0")) 
sjlabelled::set_label(df1$cleaner) <- "Know: home or office cleaner" 
#frq(df1$cleaner)

#frq(df1$v4)
df1$hairbarber <- as.numeric(car::recode(df1$v4,"1:3=1;4=0")) 
sjlabelled::set_label(df1$hairbarber) <- "Know: hairdresser/barber" 
#frq(df1$hairbarber)

#frq(df1$v5)
df1$hrmanager <- as.numeric(car::recode(df1$v5,"1:3=1;4=0")) 
sjlabelled::set_label(df1$hrmanager) <- "Know: human resource manager" 
#frq(df1$hrmanager)

#frq(df1$v6)
df1$lawyer <- as.numeric(car::recode(df1$v6,"1:3=1;4=0")) 
sjlabelled::set_label(df1$lawyer) <- "Know: lawyer" 
#frq(df1$lawyer)

#frq(df1$v7)
df1$mechanic <- as.numeric(car::recode(df1$v7,"1:3=1;4=0")) 
sjlabelled::set_label(df1$mechanic) <- "Know: car mechanic" 
#frq(df1$mechanic)

#frq(df1$v8)
df1$nurse <- as.numeric(car::recode(df1$v8,"1:3=1;4=0")) 
sjlabelled::set_label(df1$nurse) <- "Know: nurse" 
#frq(df1$mechanic)

#frq(df1$v9)
df1$police <- as.numeric(car::recode(df1$v9,"1:3=1;4=0")) 
sjlabelled::set_label(df1$police) <- "Know: police officer" 
#frq(df1$police)

#frq(df1$v10)
df1$teacher <- as.numeric(car::recode(df1$v10,"1:3=1;4=0")) 
sjlabelled::set_label(df1$teacher) <- "Know: school teacher" 
#frq(df1$teacher)

# Total of known occupations
isei_ocupnames<- names(dplyr::select(df1,busdriver:teacher))
df1$know_total <- rowSums(x = dplyr::select(df1,isei_ocupnames),na.rm = T)
summary(df1$know_total)
sjlabelled::set_label(df1$know_total) <- 'Total of known occupations'
```

```{r isei-generador}
df1$isei_busdriver <- as.numeric(ifelse(df1$busdriver ==1,37,NA)) 
sjlabelled::set_label(df1$isei_busdriver) <- "ISEI: bus/lorry driver" 
#frq(df1$isei_busdriver)

df1$isei_ceo <- as.numeric(ifelse(df1$ceo==1,70,NA)) 
sjlabelled::set_label(df1$isei_ceo) <- "ISEI: CEO large company" 
#frq(df1$isei_ceo)

df1$isei_cleaner <- as.numeric(ifelse(df1$cleaner==1,17,NA)) 
sjlabelled::set_label(df1$isei_cleaner) <- "ISEI: home or office cleaner" 
#frq(df1$isei_cleaner)

df1$isei_hairbarber <- as.numeric(ifelse(df1$hairbarber==1,32,NA)) 
sjlabelled::set_label(df1$isei_hairbarber) <- "ISEI: hairdresser/barber" 
#frq(df1$isei_hairbarber)

df1$isei_hrmanager <- as.numeric(ifelse(df1$hrmanager==1,68,NA)) 
sjlabelled::set_label(df1$isei_hrmanager) <- "ISEI: human resource manager" 
#frq(df1$isei_hrmanager)

df1$isei_lawyer <- as.numeric(ifelse(df1$lawyer==1,85,NA)) 
sjlabelled::set_label(df1$isei_lawyer) <- "ISEI: lawyer" 
#frq(df1$isei_lawyer)

df1$isei_mechanic <- as.numeric(ifelse(df1$mechanic==1,38,NA)) 
sjlabelled::set_label(df1$isei_mechanic) <- "ISEI: car mechanic" 
#frq(df1$isei_mechanic)

df1$isei_nurse <- as.numeric(ifelse(df1$nurse==1,48,NA)) 
sjlabelled::set_label(df1$isei_nurse) <- "ISEI: nurse" 
#frq(df1$isei_nurse)

df1$isei_police <- as.numeric(ifelse(df1$police==1,53,NA)) 
sjlabelled::set_label(df1$isei_police) <- "ISEI: police officer" 
#frq(df1$isei_police)

df1$isei_teacher <- as.numeric(ifelse(df1$teacher==1,63,NA)) 
sjlabelled::set_label(df1$isei_teacher) <- "ISEI: school teacher" 
#frq(df1$isei_teacher)
```

```{r social-capital-measures}
# Social capital measures
isei_ocupnames<- names(select(df1,starts_with("isei_")))

# number of high status contacts
df1$n_high <- 
  rowSums(x = select(df1,c("lawyer","hrmanager","ceo")),na.rm = T)
sjlabelled::set_label(df1$n_high) <- 'Number of high status contacts'

# number of middle status contacts
df1$n_middle <-
  rowSums(x = select(df1,c("teacher","police","nurse")),na.rm = T)
sjlabelled::set_label(df1$n_middle) <- 'Number of middle status contacts'

# number of low status contacts
df1$n_low <- 
  rowSums(x = select(df1,c("mechanic","busdriver","hairbarber","cleaner")),na.rm = T)
sjlabelled::set_label(df1$n_low) <- 'Number of low status contacts'

# Average ISEI 
#REVIEW SHOULD NOT CONTAIN 0
df1$isei_avg <- rowMeans(x = select(df1,isei_ocupnames),na.rm = T) 
sjlabelled::set_label(df1$isei_avg) <- 'Average contact prestige'
summary(df1$isei_avg)
```

- using my original coding

```{r homogeneity-R-digclass-V-working}
# EI homophily index
df1$ingroup <- NA 
df1 %>% select(dclass3_V,n_high,n_middle,n_low,ingroup)
df1$ingroup <- ifelse(df1$dclass3_V=="Service Class (I+II)",df1$n_high,df1$ingroup)
df1$ingroup <- ifelse(df1$dclass3_V=="Intermediate class (III+IV)",df1$n_middle,df1$ingroup)
df1$ingroup <- ifelse(df1$dclass3_V=="Working Class (V+VI+VII)",df1$n_low,df1$ingroup)
#frq(df1$ingroup)

df1$outgroup <- NA 
df1 %>% select(dclass3_V,n_high,n_middle,n_low,outgroup)
df1$outgroup <- ifelse(df1$dclass3_V=="Service Class (I+II)",rowSums(df1[,c("n_low","n_middle")],na.rm = T),df1$outgroup)
df1$outgroup <- ifelse(df1$dclass3_V=="Intermediate class (III+IV)",rowSums(df1[,c("n_low","n_high")],na.rm = T),df1$outgroup)
df1$outgroup <- ifelse(df1$dclass3_V=="Working Class (V+VI+VII)",rowSums(df1[,c("n_middle","n_high")],na.rm=T),df1$outgroup)
#frq(df1$outgroup)

df1$homclass3_V <- as.numeric(df1$ingroup/df1$know_total)
sjlabelled::set_label(df1$homclass3_V) <- "Class homogeneity Dominance (DIGCLASS version - V to Working)"

## Using RESPONDENT Class for homogeneity---------------------------------------
df1$ingroup_res <- NA 
df1 %>% select(dclass3res_V,n_high,n_middle,n_low,ingroup_res)
df1$ingroup_res <- ifelse(df1$dclass3res_V=="Service Class (I+II)",df1$n_high,df1$ingroup_res)
df1$ingroup_res <- ifelse(df1$dclass3res_V=="Intermediate class (III+IV)",df1$n_middle,df1$ingroup_res)
df1$ingroup_res <- ifelse(df1$dclass3res_V=="Working Class (V+VI+VII)",df1$n_low,df1$ingroup_res);frq(df1$ingroup_res)

df1$outgroup_res <- NA 
df1 %>% select(dclass3res_V,n_high,n_middle,n_low,outgroup_res)
df1$outgroup_res <- ifelse(df1$dclass3res_V=="Service Class (I+II)",rowSums(df1[,c("n_low","n_middle")],na.rm = T),df1$outgroup_res)
df1$outgroup_res <- ifelse(df1$dclass3res_V=="Intermediate class (III+IV)",rowSums(df1[,c("n_low","n_high")],na.rm = T),df1$outgroup_res)
df1$outgroup_res <- ifelse(df1$dclass3res_V=="Working Class (V+VI+VII)",rowSums(df1[,c("n_middle","n_high")],na.rm=T),df1$outgroup_res);frq(df1$outgroup_res)

df1$homclass3_V_res <- as.numeric(df1$ingroup_res/df1$know_total)
sjlabelled::set_label(df1$homclass3_V_res) <- "Class homogeneity (R's) (DIGCLASS - V to intermediate)";frq(df1$homclass3_V_res)
```

```{r homogeneity-isei, include=FALSE}
df1$dis_busdriver <- df1$isei08 - df1$isei_busdriver
df1$dis_ceo <- df1$isei08 - df1$isei_ceo
df1$dis_cleaner <- df1$isei08 - df1$isei_cleaner
df1$dis_hairbarber <- df1$isei08 - df1$isei_hairbarber
df1$dis_hrmanager <- df1$isei08 - df1$isei_hrmanager
df1$dis_lawyer <- df1$isei08 - df1$isei_lawyer
df1$dis_mechanic <- df1$isei08 - df1$isei_mechanic
df1$dis_nurse <- df1$isei08 - df1$isei_nurse
df1$dis_police <- df1$isei08 - df1$isei_police
df1$dis_teacher <- df1$isei08 - df1$isei_teacher

# 88 - 37  = 51
# 88 - 17  = 71

df1$socdist<- rowMeans(x = df1[,c("dis_busdriver","dis_ceo","dis_cleaner",
                                          "dis_hairbarber","dis_hrmanager","dis_lawyer",
                                          "dis_mechanic","dis_nurse","dis_police","dis_teacher")],na.rm = T)
sjlabelled::set_label(df1$socdist) <- "Social distance: ISEI (respondent - contacts)"
sjPlot::plot_frq(df1$socdist,type = "histogram")

df1$socdist_tot<- (df1$isei08- df1$isei_avg)
sjlabelled::set_label(df1$socdist_tot) <- "Social distance: ISEI (respondent - avg contacts)"
sjPlot::plot_frq(df1$socdist_tot,type = "histogram")

#higher values represent higher socioeconomic distance from my contacts. In other words, higher values means that in average, my contacts are more different from me in terms of socioeconomic status (ISEI), while lower values represent higher homogeneity.
```

# Demographics

```{r demographics}
df1$agenum <- as.numeric(car::recode(df1$AGE,recodes = "999=NA",as.numeric = T))
sjlabelled::set_label(df1$agenum) <- "Age in years"

df1$female <- car::recode(as.numeric(df1$SEX),"2='Female';1='Male';9=NA",
                          levels = c('Male','Female'),as.factor = T)

frq(df1$MARITAL)
df1$partner <- df1$MARITAL
frq(df1$partner)
df1$partner <- ifelse(df1$partner==9& df1$c_alphan =="CN",yes = 2,no = df1$partner)
frq(df1$partner)
df1$partner <- car::recode(df1$partner,recodes = "1:2=1;3:6=0;c(7,9)=NA",as.factor = T)
df1$partner <- factor(df1$partner,labels = c("No partner","Has a partner"))
frq(df1$partner)
```

# Save

```{r svae}
save(df1,file = here::here("input/data/proc/study1.RData"))
```
