==> processing/02-prod_prep-country.Rmd <==
---
title: "Data preparation - country"
date: "`r format(Sys.time(), '%A %d %B %Y %X')`"
output: 
  html_document: 
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: no
      number_sections: yes
    code_folding: show  
    number_sections: yes
editor_options: 
  chunk_output_type: console
---

# Setup

```{r setup}
knitr::opts_chunk$set(warning = FALSE, 
                      message = FALSE, 
                      echo = TRUE 
                      )
options(scipen=9999)
if (!require("pacman")) install.packages("pacman") 
load(here::here("input/data/proc/study1.RData"))
pacman::p_load(dplyr,sjmisc,sjPlot,sjlabelled,tidyverse,haven,MLMusingR)
```

# Country level

```{r}
countries <- 
df1 %>% 
  group_by(country) %>% 
  summarise(n=n())
countries$ccode <- as.numeric(countries$country)
countries$countryname <- as_character(countries$country)
```


```{r}
wid_gini <- haven::read_dta(here::here("input/data/original/wid_gini.dta"))
wid_ratios <- haven::read_dta(here::here("input/data/original/wid_ratios.dta"))
wid_shares <- haven::read_dta(here::here("input/data/original/wid_shares.dta"))
wid_df <-wid_ratios  %>% 
  dplyr::left_join(wid_shares,by="country_yr") %>% 
  dplyr::left_join(wid_gini,by="country_yr") 

names(wid_df)
# summary(wid_df)
wid_df <- 
wid_df %>% 
  dplyr::select(country=country.x,year=year.x,country_yr,starts_with("wid_"))

wid_df$ccode <- countrycode::countrycode(sourcevar = wid_df$country,origin = "iso2c",destination = "iso3n")

summary(wid_df)
wid_df_fill<- 
wid_df %>%
  group_by(country) %>% 
  fill(wid_p90p10, .direction = "up") %>%
  fill(wid_sp90p10, .direction = "up") %>%
  fill(wid_gini_disp, .direction = "up")
summary(wid_df_fill)
  
wid_df1 <- wid_df_fill %>% filter(year==2017) %>% select(ccode,starts_with("wid_"))
wid_df1$country <- NULL
sjPlot::tab_corr(wid_df1)
```


**World Income Inequality Database (WIID)** 

> UNU-WIDER, World Income Inequality Database (WIID) Companion dataset (wiidcountry and/or wiidglobal). Version 30 June 2022. https://doi.org/10.35188/UNU-WIDER/WIIDcomp-300622

```{r , eval=FALSE, include=FALSE}
wiid_long<- haven::read_dta(here::here("input/data/original/WIID_30JUN2022_0/WIID_30JUN2022_0.dta"))
wiid_long$ccode <- countrycode::countrycode(sourcevar = wiid_long$c3,origin = "iso3c",destination = "iso3n")

wiid_long1 <- 
wiid_long %>% 
  filter(ccode %in% countries$ccode) %>% 
  filter(year==2017)
```

```{r}
wiid <- 
  read.csv(file = here::here("input/data/original/wid-percentiles/wiid-data-2015-2020.csv"),sep = ";")
names(wiid) <- tolower(names(wiid))
names(wiid) <- gsub('\\.', '', names(wiid))
wiid$ccode <- countrycode::countrycode(sourcevar = wiid$iso,origin = "iso3c",destination = "iso3n")

wiid_d10d01 <- 
  read.csv(file = here::here("input/data/original/wid-percentiles/wiid-data-D10D01-ratio.csv"),sep = ";")

names(wiid_d10d01) <- tolower(names(wiid_d10d01))
names(wiid_d10d01) <- gsub('\\.', '', names(wiid_d10d01))
wiid_d10d01$ccode <- countrycode::countrycode(sourcevar = wiid_d10d01$iso,origin = "iso3c",destination = "iso3n")
wiid_d10d01 <- 
  wiid_d10d01 %>%  
  filter(ccode %in% countries$ccode) %>% 
  filter(year==2017) %>% 
  select(ccode,everything(),-iso,-country,-year,-giniindex) 

wiid_s80s10 <- 
  read.csv(file = here::here("input/data/original/wid-percentiles/wiid-data-P80P20.csv"),sep = ";")
names(wiid_s80s10) <- tolower(names(wiid_s80s10))
names(wiid_s80s10) <- gsub('\\.', '', names(wiid_s80s10))
wiid_s80s10$ccode <- countrycode::countrycode(sourcevar = wiid_s80s10$iso,origin = "iso3c",destination = "iso3n")

wiid_s80s10 <- 
  wiid_s80s10 %>%  
  filter(ccode %in% countries$ccode) %>% 
  filter(year==2017) %>% 
  select(ccode,everything(),-iso,-country,-year,-giniindex) 

wiid_2017 <- 
wiid %>% 
  filter(ccode %in% countries$ccode) %>% 
  filter(year==2017) %>% 
  select(ccode,everything(),-iso,-country,-year) %>% 
  left_join(wiid_d10d01,by="ccode") %>% 
  left_join(wiid_s80s10,by="ccode")
```


**Solt SWDIID**

```{r Solt}
load(file = here::here("input/data/original/swiid9_5/swiid9_5.rda"))
solt <- 
swiid_summary %>% dplyr::select(country,year,gini_disp,gini_mkt,abs_red,rel_red) %>% dplyr::filter(year %in% c(2016))
summary(solt)
solt$ccode <- countrycode::countrycode(sourcevar = solt$country,origin = "country.name",destination = "iso3n")
solt$country <- NULL
solt_2017 <-  solt %>% filter(ccode %in% countries$ccode) %>% 
  dplyr::select(ccode,gini_disp,gini_mkt)

solt_2017$abs_red <- solt_2017$gini_mkt - solt_2017$gini_disp 
solt_2017$rel_red <- solt_2017$abs_red/solt_2017$gini_mkt
summary(solt_2017)
```

**ILO Data**

```{r}
ilo_data <- read.csv(file = here::here("input/data/original/ILO-data/WEOApr2022all.csv"))
frq(ilo_data$Subject.Descriptor)

ilo_data_1 <- ilo_data %>% filter(Subject.Descriptor %in% c("General government total expenditure","General government revenue"), Units == "Percent of GDP")
ilo_data_1$ccode <- countrycode::countrycode(sourcevar = ilo_data_1$ISO,origin = "iso3c",destination = "iso3n") 
ilo_data_1<- ilo_data_1 %>% filter(ccode %in% countries$ccode)
ilo_data_1 <- ilo_data_1 %>% select(ccode,WEO.Country.Code:Country.Series.specific.Notes,X2010:X2020)
names(ilo_data_1) <- tolower(names(ilo_data_1))
names(ilo_data_1) <- str_replace_all(string = names(ilo_data_1),pattern = "x",replacement = "Y_")

ilo_tax <- ilo_data_1 %>% filter(weo.subject.code=="GGR_NGDP") 
ilo_spe <- ilo_data_1 %>% filter(weo.subject.code=="GGX_NGDP") 
# Example wide format data

# Convert from wide to long format
long_ilo_tax <- ilo_tax %>%
  select(ccode,Y_2010:Y_2016) %>% 
  pivot_longer(cols = starts_with("Y_"),
               names_to = "year",
               names_prefix = "Y_",
               values_to = "ilo_taxrev")
long_ilo_tax$ilo_taxrev <- as.numeric(long_ilo_tax$ilo_taxrev)

long_ilo_spe <- ilo_spe %>%
  select(ccode,Y_2010:Y_2016) %>% 
  pivot_longer(cols = starts_with("Y_"),
               names_to = "year",
               names_prefix = "Y_",
               values_to = "ilo_govspe")

long_ilo_spe$ilo_govspe <- as.numeric(long_ilo_spe$ilo_govspe)

ilo_df <- left_join(long_ilo_tax,long_ilo_spe) %>% filter(year==2016)
```


**GDP: PEN data**
```{r pen-data}
pendata<- haven::read_dta(file = here::here("input/data/original/pen-world/pwt1001.dta"))
pendata$ccode <- countrycode::countrycode(sourcevar = pendata$countrycode,origin = "iso3c",destination = "iso3n")
pen_2017 <- pendata %>% filter(ccode %in% countries$ccode,year %in% 2017) %>% select(ccode,everything(),-country,-countrycode,-currency_unit,-year)
```

*OCDE data*
```{r ocde}
ocde<- 
  read.csv(file = here::here("input/data/original/ocde-countries.csv"),sep = ",")

ocde$ccode <- countrycode::countrycode(sourcevar = ocde$iso3n,origin = "iso3c",destination = "iso3n")
ocde <- ocde[,c("ccode","oecd")]
```

```{r merge}
macro_data <- 
  left_join(wiid_2017,solt_2017) %>%
  left_join(pen_2017) %>% 
  left_join(ocde) %>% 
  left_join(wid_df1) %>% 
  left_join(ilo_df) 

save(macro_data,file = here::here("input/data/proc/study1_macro_data.RData"))
df1$ccode <- as.numeric(df1$country)
df2 <- left_join(df1,macro_data,by="ccode")
df2 <- 
df2 %>% mutate(country2=countrycode::countrycode(sourcevar = ccode,origin = "iso3n",destination = "iso3c"),
               country3=countrycode::countrycode(sourcevar = ccode,origin = "iso3n",destination = "country.name"),
               region=countrycode::countrycode(sourcevar = ccode,origin = "iso3n",destination = "region23")) %>% 
  select(ccode,country2,country3,everything())
```

```{r save}
save(df2,file = here::here("input/data/proc/study1_country.RData"))
sjPlot::view_df(df2,file =here::here("input/data/proc/codebook-s1.html") )
```
