==> processing/03_prod_main_analysis.Rmd <==
---
# title: "Class-based network segregation, Economic Inequality and Redistributive Preferences across societies"
fontsize: 12pt
linestretch: '1.5'          # interlineado
# #linestretch: '2.0'          # interlineado 
papersize: a4
link-citations: yes         # citas linkeadas
lang: en-US
  bookdown::html_document2:
    toc: false
    code_folding: hide
    number_sections: false
    toc_float:
      collapsed: false
      smooth_scroll: false
  bookdown::pdf_document2:
    template: null
    toc: false
    keep_tex: false
    number_sections: true
    pandoc_args:
      - --template=input/mytemplate.tex #custom template para usar autores con afiliacion  
  bookdown::word_document2:
    reference_docx: "input/mytemplate.docx"
    toc: false
    code_folding: hide
    number_sections: false
    toc_float:
      collapsed: false
      smooth_scroll: false      
         
always_allow_html: true      
linkcolor: gray                         # enlaces y citas en color azul
bibliography: input/bib/rgroup-redist.bib     # bibliografia en bibtex
csl: input/bib/apa-no-doi-no-issue.csl
editor_options:
  chunk_output_type: console            # en RStudio, mostrar output en consola
geometry: "left=2.54cm,right=2.54cm,top=2.54cm,bottom=2.54cm" # márgenes de página
header-includes:
  - \usepackage{times}   # Times New Roman
  - \usepackage{endnotes}
  - \let\footnote=\endnote 
  - \usepackage{caption}
  - \captionsetup[figure, table]{labelfont={bf},labelformat={default},labelsep=period}
  - \usepackage{graphicx}
  - \usepackage{setspace}  
  - \usepackage{float}
  - \usepackage{booktabs}
  - \usepackage{longtable}
  - \usepackage{array}
  - \usepackage{multirow}
  - \usepackage{wrapfig}
  - \usepackage{float}
  - \usepackage{colortbl}
  - \usepackage{pdflscape}
  - \usepackage{tabu}
  - \usepackage{threeparttable}
  - \setlength\parindent{24pt} 
  - \usepackage{fancyhdr}
  - \fancyhead{} 
---

```{r setup, include=FALSE}
if (!require("pacman")) install.packages("pacman")  #si falta pacman, instalar
if (!require("tinytex")) install.packages("tinytex")#si falta tinytex, instalar
pacman::p_load(knitr, kableExtra, dplyr, ggplot2,sjmisc,texreg) # librerias
knitr::opts_chunk$set(warning = FALSE,  # mensaje de warning
                      message = FALSE,  # mensajes/avisos de librerias  
                      cache = T,    # cache de los chunks,usar analisis pesados
                      out.width = '100%',
                      fig.pos= "H",     # posicion figuras H = HERE
                      fig.align = "center",
                      echo = FALSE      # incluir chunk en output
                      )
options(scipen=999) # notacion cientifica
rm(list=ls())       # limpiar workspace
options(knitr.kable.NA = '') # NA en kable = ''  
```

# Results

## Descriptive cross-country comparison on class, network segregation, and redistributive preferences

```{r fig2, fig.dim=c(10,5),fig.cap='Cross-country differences in redistributive preferences by social class',warning=FALSE}
set1 <-  RColorBrewer::brewer.pal(n = 8, name = "Set1")
options(ggplot2.discrete.fill = set1) 
options(ggplot2.discrete.colour = set1) 
ggplot2::theme_set(ggplot2::theme_bw()) 
ggplot2::theme_update(text=ggplot2::element_text(size=15,  family="serif"))
options(scipen=999)
if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,ggplot2,haven,ggrepel,gridExtra)  
rm(list=ls())
load(here::here("input/data/proc/study1_country.RData"));df1 <- df2

dfreg <- df1 %>% dplyr::select(
  redist,prefineq,
  egal = egal2,
  dclass3res_V,
  homclass3_V_res,
  know_total,
  Q03pcm,
  edyears,
  female,
  agenum,
  partner,
  workst,
  WEIGHT,
  region,
  "gini_disp"=wid_gini_disp,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  gdppercapita,
  country2, country, country3,
) %>%
  filter(country2 != "SVN") %>% 
  na.omit() %>% 
    mutate(loggdppercapita=scale(gdppercapita/1000))

df_mean <- 
dfreg %>% 
  group_by(country3) %>% 
  summarise(egal2_c=weighted.mean(egal,WEIGHT),homclass_c=weighted.mean(homclass3_V_res,WEIGHT),gini_disp=mean(gini_disp))
df_region <- dfreg %>% group_by(region,country3) %>% summarise(n())
df_full <- dfreg %>% group_by(dclass3res_V) %>%
  summarise(egal2=weighted.mean(egal,WEIGHT),
            homclass=weighted.mean(homclass3_V_res,WEIGHT),
            homclass_c=weighted.mean(dfreg$homclass3_V_res,dfreg$WEIGHT),
            egal2_c=weighted.mean(dfreg$egal,dfreg$WEIGHT), 
            country3="Sample")
df_plot <- dfreg %>% group_by(country3,dclass3res_V) %>% 
  summarise(egal2=weighted.mean(egal,WEIGHT),
            homclass=weighted.mean(homclass3_V_res, WEIGHT)) %>% 
  left_join(df_mean) %>% left_join(df_region) 
df_plot$dclass3res_V <- factor(df_plot$dclass3res_V,levels = levels(df_plot$dclass3res_V),labels = c("Service Class","Intermediate Class","Working Class"))

# Calculate the difference between the mean value of working class and upper class within each country
homo_class_diff <- df_plot %>%
  dplyr::select(country3,homclass,dclass3res_V) %>% 
  tidyr::pivot_wider(names_from = "dclass3res_V", values_from = "homclass") %>% 
  mutate(diff_working_upper = `Working Class` - `Service Class`) %>% 
  dplyr::select(-`Working Class`, -`Service Class`,-`Intermediate Class`) %>% 
  as.data.frame()

egal_class_diff <- df_plot %>%
  dplyr::select(country3,egal2,dclass3res_V) %>% 
  tidyr::pivot_wider(names_from = "dclass3res_V", values_from = "egal2") %>% 
  mutate(diff_working_upper_egal = `Working Class` - `Service Class`) %>% 
  dplyr::select(-`Working Class`, -`Service Class`,-`Intermediate Class`) %>% 
  as.data.frame()

df_plot <- 
dfreg %>% 
  group_by(country3,dclass3res_V) %>% 
  summarise(egal2=weighted.mean(egal,WEIGHT),homclass=weighted.mean(homclass3_V_res,WEIGHT)) %>% 
  left_join(df_mean) %>%
  left_join(df_region) %>% as.data.frame() %>% 
  left_join(homo_class_diff) %>% 
  left_join(egal_class_diff)

df_plot2 <- df_plot
df_plot2$dclass3res_V <- "Country"
df_plot2$homclass <- df_plot2$homclass_c
df_plot2$egal2 <- df_plot2$egal2_c 
df_plot3 <- bind_rows(df_plot,df_plot2)
df_plot3$dclass3res_V <- factor(df_plot3$dclass3res_V,
                                levels = c("Country","Service Class (I+II)","Intermediate class (III+IV)","Working Class (V+VI+VII)"),labels = c("Country","Service Class","Intermediate Class","Working Class"))

df_plot3 %>% 
ggplot()+
  geom_hline(yintercept = mean(dfreg$egal),linetype=2)+
  geom_point(aes(x = reorder(country3, egal2_c),y = egal2,
                 shape=dclass3res_V,
                 color=dclass3res_V
                 ),size=3, alpha=2) +
  scale_shape_manual(values = c(4,17, 15, 16)) +
  scale_color_manual(values = c("black","#000000","#666666","#999999")) +
  labs(x=NULL,y="Redistributive Preferences",caption="Source: International Social Survey 2017; descriptive statistics; figure report country averages in redistributive preferences by social class") +
  theme(legend.position = "bottom",
        legend.direction = "horizontal",
        axis.text.x=element_text(angle=45,hjust = 1,size=8),
        legend.title = element_blank())

ggsave(plot = last_plot(),filename = "figure002.png",device = "tiff",
       path = here::here("output/images"),width = 2,height = 1,units = "cm",scale = 15,dpi = 500)
```


```{r fig3, fig.dim=c(10,5),fig.cap='Cross-country differences in network homogeneity by social class ',warning=FALSE}
df_plot3 %>%
ggplot()+
  geom_point(aes(x = reorder(country3, diff_working_upper),
                 y = homclass,
                 shape=dclass3res_V,
                 color=dclass3res_V),
             size=3, alpha=2) +
  geom_hline(yintercept = mean(dfreg$homclass3_V_res),linetype=2)+
  scale_shape_manual(values = c(4,17, 15, 16)) +
  scale_color_manual(values = c("black","#000000","#666666","#999999")) +
  scale_y_continuous(limits = c(0,0.7))+
  labs(x=NULL,y="Network homogeneity",caption="Source: International Social Survey 2017; descriptive statistics; figure report country averages in network homogeneity by social class") +
  theme(legend.position = "bottom",
        legend.direction = "horizontal",
        axis.text.x=element_text(angle=45,hjust = 1,size=8),
        legend.title = element_blank())  
ggsave(plot = last_plot(),filename = "figure003.png",device = "tiff",
       path = here::here("output/images"),width = 2,height = 1,units = "cm",scale = 15,dpi = 500)
```


```{r fig4, fig.dim=c(10,5),fig.cap='Relationship between income Inequality, network homogeneity and redistributive preferences',warning=FALSE}
set1 <-  RColorBrewer::brewer.pal(n = 8, name = "Set1")
options(ggplot2.discrete.fill = set1)
options(ggplot2.discrete.colour = set1)
ggplot2::theme_set(ggplot2::theme_bw())
ggplot2::theme_update(text=ggplot2::element_text(size=15,  family="serif"))    
options(scipen=999)
if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,ggplot2,haven,ggrepel,gridExtra,grid)  
rm(list=ls())

load(here::here("input/data/proc/study1_country.RData"));df1 <- df2

dfreg <- df1 %>% dplyr::select(
  redist,prefineq,
  egal = egal2,
  dclass3res_V,
  # dclass3spo_V,
  homclass3_V_res,
  know_total,
  # socialtrust,
  Q03pcm,
  edyears,
  female,
  agenum,
  # religion,
  partner,
  # union,
  workst,
  WEIGHT,
  region,
  "gini_disp"=wid_gini_disp,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  gdppercapita,
  country2, country, country3,
) %>%
  # filter(oecd == "OECD") %>%
  filter(country2 != "SVN") %>% 
  na.omit() %>% 
    mutate(loggdppercapita=scale(gdppercapita/1000),
         edyears2 = edyears ^ 2,
         age2 = agenum ^ 2, 
         # gini_disp=scale(gini_disp),
         )

homo_ineq <- dfreg %>%
  group_by(country2,country3) %>%
  summarise_at(c("homclass3_V_res","egal","gini_disp"), mean, na.rm = TRUE) %>% 
  as.data.frame()

homo_class_diff <- dfreg %>%   
  group_by(country2,country3,dclass3res_V) %>%
  summarise_at(c("egal","homclass3_V_res"), mean, na.rm = TRUE) %>% 
  as.data.frame() %>%
  dplyr::select(country3,homclass3_V_res,dclass3res_V) %>% 
  tidyr::pivot_wider(names_from = "dclass3res_V", values_from = "homclass3_V_res") %>% 
  mutate(diff_working_upper_homo = `Working Class (V+VI+VII)` - `Service Class (I+II)`) %>% 
  as.data.frame()

# Dataset with class differences in network homogeneity
homo_ineq <- homo_ineq %>% left_join(homo_class_diff)

# Plot Homogeneity and redistributive preferences
homo_ineq$country <- countrycode::countrycode(sourcevar = homo_ineq$country2,origin = "iso3c",destination = "iso2c")

fig_homodiff_gini<- 
homo_ineq %>%  
ggplot(aes(y=diff_working_upper_homo, x=gini_disp)) +
  geom_smooth(method = "lm",fullrange=TRUE, color="black",linetype = 2,size=1,se = T) +
  # geom_text() +
  geom_text_repel(label=as.character(homo_ineq$country)) +
  xlab("Income Inequality") +
  ylab("Class Diff. in homogeneity (Working - Service)") +
  labs(caption =paste("r = ",round(cor(homo_ineq$diff_working_upper_homo,
                                     homo_ineq$gini_disp, use='complete.obs'),digits = 2)),tag = "C")


fig_homodiff_redist<- 
homo_ineq %>%  
ggplot(aes(y=egal , x=diff_working_upper_homo)) +
  geom_smooth(method = "lm",fullrange=TRUE, color="black",linetype = 2,size=1,se = T) +
  # geom_text() +
  geom_text_repel(label=as.character(homo_ineq$country)) +
  ylab("Redistributive preferences") +
  xlab("Class Diff. in homogeneity (Working - Service)") +
  labs(caption =paste("r = ",round(cor(homo_ineq$diff_working_upper_homo,
                                     homo_ineq$egal, use='complete.obs'),digits = 2)),tag = "D")

fig_egal_homo <- 
homo_ineq %>%  
ggplot(aes(y=egal, x=homclass3_V_res)) +
  geom_smooth(method = "lm",fullrange=TRUE, color="black",linetype = 2,size=1,se = T) +
  geom_text_repel(label=as.character(homo_ineq$country)) +
  xlab("Network homogeneity")+
  ylab("Redistributive preferences") +
  labs(caption =paste("r = ",round(cor(homo_ineq$homclass3_V_res,
                                     homo_ineq$egal, use='complete.obs'),digits = 2)),tag = "B")
fig_homo_gini<- 
homo_ineq %>%  
ggplot(aes(y=homclass3_V_res, x=gini_disp)) +
  geom_smooth(method = "lm",fullrange=TRUE, color="black",linetype = 2,size=1,se = T) +
  # geom_text() +
  geom_text_repel(label=as.character(homo_ineq$country)) +
  xlab("Income Inequality") +
  ylab("Network homogeneity") +
  labs(caption =paste("r = ",round(cor(homo_ineq$homclass3_V_res,
                                     homo_ineq$gini_disp, use='complete.obs'),digits = 2)),tag = "A")


set.seed(1234)
ggsave(plot = arrangeGrob(fig_homo_gini,fig_egal_homo,
                          fig_homodiff_gini,fig_homodiff_redist,
             nrow = 2, 
 bottom =textGrob("Source: International Social Survey 2017; descriptive statistics; figure report country level pearson correlation; the line is the fitted values; CI at 95%",
 gp = gpar(fontface = 1, fontsize = 9),hjust = 1,x = 1
  )),
 filename = "figure004.png",device = "tiff",
 path = here::here("output/images"),width = 1,height = 1,units = "cm",scale = 26,dpi = 500)
```

## The segregation hypothesis on redistributive preferences

```{r table1, results='asis'}
set1 <-  RColorBrewer::brewer.pal(n = 8, name = "Set1")
options(ggplot2.discrete.fill = set1)
options(ggplot2.discrete.colour = set1)
ggplot2::theme_set(ggplot2::theme_bw())
ggplot2::theme_update(text=ggplot2::element_text(size=15,  family="serif"))

if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,sjmisc,ggplot2,interplot,marginaleffects,sjlabelled,haven,stringr,ordinal,texreg,MLMusingR,lme4)  
rm(list=ls())

load(here::here("input/data/proc/study1_country.RData"));df1 <- df2
dfreg <- df1 %>% dplyr::select(
  prefineq,redist,
  egal = egal2,
  dclass3res_V,
  # dclass3spo_V,
  homclass3_V_res,
  know_total,
  # socialtrust,
  # Q03pcm,
  # edyears,
  female,
  agenum,
  # partner,
  # religion,
  # union,
  # workst, 
  WEIGHT,
  region,
  "gini_disp"=wid_gini_disp,
  "gini_mkt",
  rel_red,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  rgdpna,
  gdppercapita,
  oecd,
  country2, country, country3,
) %>%
  mutate(logrgdpna = log(rgdpna),
         loggdppercapita=log(gdppercapita)
         ) %>%
  # filter(oecd == "OECD") %>%
  filter(country2 != "SVN") 
dfreg$egal <- as.numeric(scale(dfreg$egal))
dfreg <- na.omit(dfreg)

## Model -  EGP3 - DIGCLASS - V to Working -----------------------------$ 
digclas1_VW_0<- lmer(egal~1 + dclass3res_V + female+agenum+know_total+(1|country2),data=dfreg,weights = WEIGHT)
# digclas1_VW_1<- lmer(egal~1 + homclass3_V_res+know_total +female+agenum+(1|country2),data=dfreg,weights = WEIGHT)
# digclas1_VW_2<- update(digclas1_VW_1, . ~ . +know_total) 
digclas1_VW_3 <- update(digclas1_VW_0, . ~ . + dclass3res_V)
# digclas1_VW_5 <- update(digclas1_VW_3, . ~ . + edyears+ Q03pcm+workst)

digclas1_VW_7 <- update(digclas1_VW_0, . ~ . +dclass3res_V*homclass3_V_res+know_total)
digclas1_VW <- list(digclas1_VW_0,
                    # digclas1_VW_1,
                    # digclas1_VW_3,
                    # digclas1_VW_4,
                    #digclas1_VW_5,
                    # digclas1_VW_6,
                    digclas1_VW_7)

ccoef <- list("(Intercept)"= "(Intercept)",
  "dclass3res_VIntermediate class (III+IV)"="Intermediate Class",
              "dclass3res_VWorking Class (V+VI+VII)"="Working Class",
              homclass3_V_res="Class-based network homogeneity",
              "dclass3res_VIntermediate class (III+IV):homclass3_V_res" = "Intermediate Class × Homogeneity",
              "dclass3res_VWorking Class (V+VI+VII):homclass3_V_res" = "Working Class × Homogeneity"
)

screenreg(digclas1_VW)

htmlreg(digclas1_VW,stars = c(0.001, 0.01, 0.05, 0.1),single.row = T,symbol = ".",
       caption = paste("Table 1: Multilevel models for social class, network homogeneity and redistributive preferences"),
       custom.note = "Note: Standard errors in parentheses. %stars",
       threeparttable = T,
       caption.above = T,leading.zero = T,
       custom.coef.map = ccoef,
       groups =
         list("Social Class (Ref.= Service Class)" = 2:3,
              "Social Class × Homogeneity"=5:6),
      custom.gof.rows = list("Controls"=c(rep("Yes",2))),include.loglik = FALSE,include.aic = FALSE,
      custom.gof.names = c(NA,NA,"Num. groups","Var: Country (Intercept)","Var: Residual"),
      file = here::here("output/tables/table001.html")
      )
```

```{r fig5, fig.dim=c(10,5),fig.cap='Interaction of network homogeneity and social class on redistributive preferences',warning=FALSE}
int_homo <- digclas1_VW_7
df_pred1 <- 
  predictions(int_homo, condition = c("homclass3_V_res","dclass3res_V"),
              newdata = datagrid(homclass3_V_res = seq(min(dfreg$homclass3_V_res),
                                                       max(dfreg$homclass3_V_res), by= 0.1),
                                 dclass3res_V = levels(dfreg$dclass3res_V),
                                 know_total = mean(dfreg$know_total)))
# names(df_pred1) 
df_pred1 %>% 
ggplot(aes(y =estimate,x=homclass3_V_res, fill=dclass3res_V, color=dclass3res_V,ymin=conf.low, ymax=conf.high)) +
  geom_ribbon(alpha=0.8,size=1,linetype="solid",color="black") +
  geom_line(size=1,color="black") +
  geom_point(size=2,color="black") +
  facet_grid(~factor(dclass3res_V, levels=
                       c("Working Class (V+VI+VII)","Intermediate class (III+IV)","Service Class (I+II)"),
                     c("Working Class","Intermediate Class","Service Class"))) +
  scale_fill_manual(values=c("#323232","#989898","#CCCCCC"))+
  scale_color_manual(values=c("#323232","#989898","#CCCCCC"))+
  scale_y_continuous(limits = c(-0.5,0.5))+
  labs(caption = paste0("Note: N (individuals) = ",nobs(int_homo),"; N (countries) = ",lme4::ngrps(int_homo),". Predictive estimates with 95% confidence intervals"),
       x="Network homogeneity",y="Redistributive preferences (z-score)") +
  theme(legend.position = "none", legend.title = element_blank())

ggsave(plot = last_plot(),filename = "figure005.png",device = "tiff",
       path = here::here("output/images"),width = 2,height = 1,units = "cm",scale = 15,dpi = 500)
```

## The mitigation hypothesis on network segregation and redistributive preferences

```{r table2, results='asis', cache=TRUE}
set1 <-  RColorBrewer::brewer.pal(n = 8, name = "Set1")
options(ggplot2.discrete.fill = set1)
options(ggplot2.discrete.colour = set1)
ggplot2::theme_set(ggplot2::theme_bw())
ggplot2::theme_update(text=ggplot2::element_text(size=15,  family="serif"))
options(scipen=999)
if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,sjmisc,ggplot2,interplot,marginaleffects,sjlabelled,haven,stringr,ordinal,texreg,MLMusingR,lme4)  
rm(list=ls()) 

load(here::here("input/data/proc/study1_country.RData"));df1 <- df2

dfreg <- df1 %>% dplyr::select(
  redist,prefineq,
  egal = egal2,
  dclass3res_V,
  homclass3_V_res,
  know_total,
  # Q03pcm,
  # edyears,
  female,
  agenum,
  # partner,
  # workst,
  WEIGHT,
  region,
  "gini_disp"=wid_gini_disp,
  "gini_mkt",
  rel_red,
  abs_red,
  ilo_taxrev,
  ilo_govspe,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  rgdpna,
  gdppercapita,
  oecd,
  country2, country, country3,
) %>%
  filter(country2 != "SVN") %>% 
  na.omit() %>% 
    mutate(loggdppercapita=scale(gdppercapita/1000),
         age2 = agenum ^ 2, 
         gini_disp=scale(gini_disp),
         d10d1=scale(d10d1),
         wid_p90p50=scale(wid_p90p50)
         )

fit_gov<- psych::pca(dfreg[,c("ilo_taxrev","ilo_govspe","abs_red")])
dfreg$wstate<- as.numeric(fit_gov$scores)
dfreg$wstate2 <- ((dfreg$wstate-min(dfreg$wstate,na.rm = T))/(max(dfreg$wstate,na.rm = T)-min(dfreg$wstate,na.rm = T)))*100
dfreg$wstate2 <- as.numeric(scale(dfreg$wstate2))
dfreg$egal <- as.numeric(scale(dfreg$egal))

# Variable Group-Centering -----------------------------------------------------
dfreg <- 
  dfreg %>% 
  mutate(to_dummy(female)
  )
dfreg$female_gc = group_center(dfreg$female_2, grp = dfreg$country2)
dfreg$agenum_gc = group_center(dfreg$agenum, grp = dfreg$country2)
dfreg$homclass3_V_gc <-group_center(dfreg$homclass3_V_res, grp = dfreg$country2)
dfreg$know_total_gc = group_center(dfreg$know_total, grp = dfreg$country2)
# drop dummies
dfreg <- dplyr::select(dfreg,-c(female_1,female_2))

## Cross-level interactions-----------------------------------------------------
homo_V_gini <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         homclass3_V_gc*dclass3res_V+
         female_gc+agenum_gc+
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = NULL)

homo_V_gini_int <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         gini_disp +
         homclass3_V_gc*
         dclass3res_V*
         gini_disp +
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = NULL)

digclas1_macro <- list(homo_V_gini,homo_V_gini_int)
screenreg(digclas1_macro)

ccoef <- list("dclass3res_VIntermediate class (III+IV)"="Intermediate Class",
              "dclass3res_VWorking Class (V+VI+VII)"="Working Class",
              homclass3_V_gc="Class-based network homogeneity",
              "homclass3_V_gc:dclass3res_VIntermediate class (III+IV)" = "Intermediate Class × Homogeneity",
              "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII)" = "Working Class × Homogeneity",
              "gini_disp" = "Income inequality (Gini index)",
              loggdppercapita = "GDP/capita",
              wstate2="Size of the welfare state",
              "homclass3_V_gc:gini_disp"= "Homogeneity × Income Inequality",
              "dclass3res_VIntermediate class (III+IV):gini_disp" = "Intermediate Class × Income Inequality",
              "dclass3res_VWorking Class (V+VI+VII):gini_disp" = "Working Class × Income Inequality",
              "homclass3_V_gc:dclass3res_VIntermediate class (III+IV):gini_disp" = "Intermediate Class × Homogeneity × Income Inequality",
              "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII):gini_disp" = "Working Class × Homogeneity × Income Inequality")

htmlreg(digclas1_macro,stars = c(0.001, 0.01, 0.05, 0.1),scalebox = 0.7,symbol = ".",
       caption = paste("Table 2: Multilevel models for income inequality, social class, network homogeneity and redistributive preferences"),
       custom.note = " Note: Individual level variables are centered within cluster (group mean).Standard errors in parentheses. %stars",
       caption.above = T,leading.zero = T,single.row = T,threeparttable = T,
       custom.coef.map = ccoef,
      groups = list("Social Class (Ref.= Service Class)" = 1:2,
                    "Social Class × Homogeneity"=4:5,
                    "Class × Income Inequality" = 8:9 ,
                    "Social Class × Homogeneity × Income Inequality"=10:11),
      custom.gof.rows = list("Controls"=rep("Yes",2)),
      include.loglik = FALSE,include.aic = FALSE,
      custom.gof.names = c(NA,NA,"Num. groups",
                           "Var: Country (Intercept)",
                           "Var: Country Homogeneity",
                           "Var: Country Intermediate Class ",
                           "Var: Country Working Class",
                           "Cov: Country (Intercept), Homogeneity",
                           "Cov: Country (Intercept), Intermediate Class ",
                           "Cov: Country (Intercept), Working Class",
                           "Cov: Country Homogeneity, Intermediate Class",
                           "Cov: Country Homogeneity, Working Class",
                           "Cov: Country Intermediate Class, Working Class",
                           "Var: Residual"),
        file = here::here("output/tables/table002.html"))
```

```{r fig6, fig.dim=c(12,7),fig.cap='Three-way interaction of network homogeneity, social class and income inequality on redistributive preferences'}
### Predicted values homo*class*GiniD ---------------------------------------$
ginid_max<- round(mean(dfreg$gini_disp) + sd(dfreg$gini_disp)  ,2)
ginid_min<- round(mean(dfreg$gini_disp) -sd(dfreg$gini_disp) ,2)
ginid_mea<- round(mean(dfreg$gini_disp),2)
 
df_pred_giniD_gc <-
  predictions(
    homo_V_gini_int,newdata = datagrid(
      gini_disp = c(ginid_min, ginid_mea, ginid_max),
      wstate2=mean(dfreg$wstate2),
      loggdppercapita=mean(dfreg$loggdppercapita),
      homclass3_V_gc = seq(min(dfreg$homclass3_V_gc)+0.01, max(dfreg$homclass3_V_gc)+0.01, by=0.1),
      dclass3res_V=levels(dfreg$dclass3res_V)
    )
  )
df_pred_giniD_gc <- as.data.frame(df_pred_giniD_gc)
df_pred_giniD_gc$gini_disp <- factor(df_pred_giniD_gc$gini_disp,labels = c("Gini (-1SD)","Gini (Mean)", "Gini (+1SD)"))
df_pred_giniD_gc <- df_pred_giniD_gc %>% filter(dclass3res_V!="Intermediate class (III+IV)")

df_pred_giniD_gc %>% 
  ggplot(aes(y=estimate,x=homclass3_V_gc, fill=dclass3res_V,color=dclass3res_V,ymin=conf.low, ymax=conf.high)) +
  geom_ribbon(alpha=0.8,size=0.7,linetype="solid",color="black") +
  geom_line(size=1,color="black") +
  geom_point(size=3, shape = 21,color="black") +
  facet_wrap(~gini_disp) +
  labs(y = "Redistributive preferences (z-score)",
       x="Network homogeneity",
       caption = paste0("Note: N (Individuals) = ",nobs(homo_V_gini_int),"; N (Countries) = ",lme4::ngrps(homo_V_gini_int),
                        ". Predictive estimates with 95% confidence intervals")) +
  scale_x_continuous(sec.axis = sec_axis(~ . , name = "Income inequality (post-tax) \n ", breaks = NULL, labels = NULL),
                     ) +
 
   scale_fill_manual(values=c("#323232","#CCCCCC"))+
  scale_color_manual(values=c("#323232","#CCCCCC"))+
  theme(legend.position = "bottom",
        legend.direction = "horizontal",
        axis.text=element_text(size=15),
        legend.title = element_blank(),
        legend.background = element_blank(),
        legend.box.background = element_rect(colour = "black"))

ggsave(plot = last_plot(),filename = "figure006.png",device = "tiff",
       path = here::here("output/images"),width = 2,height = 1,units = "cm",scale = 15,dpi = 500)
```

```{r tableA1, results='asis'}
set1 <-  RColorBrewer::brewer.pal(n = 8, name = "Set1")
options(ggplot2.discrete.fill = set1)
options(ggplot2.discrete.colour = set1)
ggplot2::theme_set(ggplot2::theme_bw())
ggplot2::theme_update(text=ggplot2::element_text(size=15,  family="serif"))

if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,sjmisc,ggplot2,interplot,marginaleffects,sjlabelled,haven,stringr,ordinal,texreg,MLMusingR,lme4)  
rm(list=ls())

load(here::here("input/data/proc/study1_country.RData"));df1 <- df2
dfreg <- df1 %>% dplyr::select(
   prefineq,redist,
  egal = egal2,
  dclass3res_V,
  homclass3_V_res,
  know_total,
  female,
  agenum,
  Q03pcm,
  edyears,
  partner,
  workst,
  WEIGHT,
  region,
  "gini_disp"=wid_gini_disp,
  "gini_mkt",
  rel_red,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  rgdpna,
  gdppercapita,
  oecd,
  country2, country, country3,
) %>%
  mutate(logrgdpna = log(rgdpna),
         loggdppercapita=log(gdppercapita)
         ) %>%
  filter(country2 != "SVN") 
dfreg$egal <- as.numeric(scale(dfreg$egal))
dfreg_1 <- na.omit(dfreg %>% dplyr::select(prefineq,redist,
  egal,
  dclass3res_V,
  homclass3_V_res,
  know_total,
  female,
  agenum,
  WEIGHT,
  region,country2)
  )

## Model -  EGP3 - DIGCLASS - V to Working -----------------------------$ 
digclas1_VW_0<- lmer(egal~1 + (1|country2),data=dfreg_1,weights = WEIGHT)
digclas1_VW_2 <- update(digclas1_VW_0, . ~ . + dclass3res_V)
digclas1_VW_3 <- update(digclas1_VW_2, . ~ . + dclass3res_V+homclass3_V_res+female+agenum+know_total)
digclas1_VW_7 <- update(digclas1_VW_3, . ~ . +dclass3res_V*homclass3_V_res)
#Include additional controls
digclas1_VW_8 <- update(digclas1_VW_7, . ~ . + edyears+Q03pcm+partner+workst,data=na.omit(dfreg))
digclas1_VW <- list(digclas1_VW_2,digclas1_VW_3,digclas1_VW_7,digclas1_VW_8)

ccoef <- list("dclass3res_VIntermediate class (III+IV)"="Intermediate Class",
             "dclass3res_VWorking Class (V+VI+VII)"="Working Class",
              homclass3_V_res="Class-based network homogeneity",
              know_total="Network size",
              "dclass3res_VIntermediate class (III+IV):homclass3_V_res" = "Intermediate Class × Homogeneity",
              "dclass3res_VWorking Class (V+VI+VII):homclass3_V_res" = "Working Class × Homogeneity",
             "edyears"="Year of Education",
             "Q03pcmT02"="Income (T2)","Q03pcmT03"="Income (T3)","Q03pcmMissing"="Income (No information)",
             "workstNot in paid work" = "Not in paid work (Ref. = In paid work)",
             "partnerHas a partner" = "Has partner (Ref.= No partner)",
             "femaleFemale"="Female (Ref. = Male)",
             "agenum"="Age"
             )

htmlreg(digclas1_VW,stars = c(0.001, 0.01, 0.05, 0.1),single.row = T,symbol = ".",
       caption = paste("Table A1: Multilevel models for network homogeneity and redistributive preferences"),
       custom.note = "Note: Standard errors in parentheses. %stars",
       threeparttable = T,
       caption.above = T,leading.zero = T,
       custom.coef.map = ccoef,
      groups = list("Social Class (Ref.= Service Class)" = 1:2,
                    "Social Class × Homogeneity"=5:6,
                    "Household Income (Ref.= Tertile I)"= 8:10
                    ),
      custom.gof.rows = list("Controls"=c(rep("Yes",4))),include.loglik = FALSE,include.aic = FALSE,
      custom.gof.names = c(NA,NA,"Num. groups","Var: Country (Intercept)","Var: Residual"),
      file = here::here("output/tables/tableA1.html")
      )
```

```{r tableA2, echo=FALSE, cache=TRUE, results='asis'}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,sjmisc,sjlabelled,haven,vtable)  
rm(list=ls())
load(here::here("input/data/proc/study1_country.RData"));df1 <- df2
  pacman::p_load(knitr)
tabformat = if(is_html_output()) {"html"} else if(is_latex_output()) {"latex"}
df1$id <- 1:dim(df1)[[1]]
dfreg <- df1 %>% dplyr::select(id,
  redist,prefineq,
  egal = egal2,
  dclass3res_V,dclass6res,
  homclass3_V_res,
  know_total,
  # Q03pcm,
  # edyears,
  female,
  agenum,
  # partner,
  # workst,
  "gini_disp"=wid_gini_disp,
  abs_red,
  ilo_taxrev,
  ilo_govspe,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  gdppercapita,
  country2,country3,
) %>%
  filter(country2 != "SVN") %>% 
  na.omit() %>%
    mutate(loggdppercapita=gdppercapita/1000,
         age2 = agenum ^ 2, 
         )

occup_prop <- df1 %>% 
  filter(id %in% dfreg$id) %>% 
  dplyr::select(hrmanager,ceo,lawyer,
         nurse,police,teacher,
         cleaner,hairbarber,busdriver,mechanic
         ) %>% summarise_all(mean,na.rm=T)

p <- t(round(occup_prop*100,1))[,1]
pacman::p_load(dplyr,knitr,kableExtra)  
# Create the data frame
df <- data.frame(
  # Item = c("c", "d", "a", "g", "h", "i", "j", "e", "b", "f"),
  Occupation = c("Home or office cleaner", "Hairdresser/barber", "Bus/lory driver", "Car mechanic", 
                 "Nurse", "Police officer", "School teacher", "Human resource manager", 
                 "Executive of large firm", "Lawyer"),
  "ISCO08" = rev(c(2611,1100,1212,2300,5412,2220,7231,8331,5140,9111))
) %>% 
  mutate("%"=p)
tab_an_posgen <- 
kable(df,linesep = "",booktabs=TRUE,format=tabformat,caption = "Table A2: Occupations included in the position generator instrument") %>%
  kable_classic(full_width = F,font_size = 15,latex_options = "HOLD_position",html_font = "sans") %>%
  column_spec(1,width = "6cm") %>%
  pack_rows("Higher-status positions", 1, 3) %>%
  column_spec(2) %>% 
  pack_rows("Medium-status positions", 4, 6) %>% 
  pack_rows("Lower-status positions", 7, 10) 
tab_an_posgen

kableExtra::save_kable(tab_an_posgen, here::here("output/tables/tableA02.html"))
```

```{r tableA3,results='asis', cache=TRUE}
levels(dfreg$dclass6res) <- c("Upper Service","Lower Service","Routine nonmanual",
                      "Self-employed","Skilled working","Unskilled working")
levels(dfreg$dclass3res_V) <- c("Service","Intermediate","Working")

tab_an_classag <- 
left_join(
dfreg %>% 
  group_by(dclass6res,dclass3res_V) %>% summarise(N1=n()) %>% mutate(freq1 = (N1/ dim(dfreg)[[1]])*100),
dfreg %>% 
  group_by(dclass3res_V) %>% summarise(N2=n()) %>% mutate(freq2 = (N2 / dim(dfreg)[[1]])*100),by="dclass3res_V"
) %>% 
  dplyr::select("dclass6res","N1","freq1","dclass3res_V","N2","freq2") %>%
  kable(booktabs = TRUE, format=tabformat,linesep = "",digits = 1,format.args = list(big.mark = ","),
        caption = "Table A3: Level of aggregation of social class",
        col.names = c("EGP-6","N","%","EGP-3","N","%")) %>% 
  kable_classic(full_width = F,font_size = 15,latex_options = c("HOLD_position"),html_font = "sans") %>% 
  collapse_rows(columns = c(4,5,6), valign = "middle");tab_an_classag

kableExtra::save_kable(tab_an_classag, here::here("output/tables/tableA03.html"))

```

```{r tableA4, echo=FALSE, results='asis'}
dfreg <- df1 %>% dplyr::select(
  redist,prefineq,
  egal = egal2,
  dclass3res_V,dclass6res,
  homclass3_V_res,
  know_total,
  # Q03pcm,
  # edyears,
  female,
  agenum,
  # partner,
  # workst,
  "gini_disp"=wid_gini_disp,
  abs_red,
  ilo_taxrev,
  ilo_govspe,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  gdppercapita,
  country2,country3,country,
) %>%
  filter(country2 != "SVN") %>% 
  na.omit() %>% 
    mutate(loggdppercapita=gdppercapita/1000)

fit_gov<- psych::pca(dfreg[,c("ilo_taxrev","ilo_govspe","abs_red")])
dfreg$wstate<- as.numeric(fit_gov$scores)
dfreg$wstate2 <- ((dfreg$wstate-min(dfreg$wstate,na.rm = T))/(max(dfreg$wstate,na.rm = T)-min(dfreg$wstate,na.rm = T)))*100


dfreg$country <- countrycode::countrycode(sourcevar = dfreg$country2,origin = "iso3c",destination = "iso2c")
dfreg$country3b <- paste0(dfreg$country3," (",dfreg$country,")")

tab_an_countries <- 
dfreg %>% 
  group_by(country3b) %>% 
  summarise(n=n(),
            homclass=mean(x=homclass3_V_res),
            gini_disp=mean(gini_disp),
            loggdppercapita=mean(loggdppercapita),
            wstate=mean(wstate2)) %>% 
  arrange(homclass) %>% ungroup() %>% 
  kable(digits = c(1,1,3,2,2,2),booktabs = TRUE,format = tabformat, linesep = "",format.args = list(big.mark = ","),align = "lrccccc",
        caption = "Table A4: Values per country for the country-level variables",
        col.names = c("Country","N","Network Homogeneity","Income Inequality (Gini Index)","GDP/capita in $1000","Size of the Welfare State")) %>% 
  kable_classic(full_width = F,latex_options = "scale_down",html_font = "sans") %>% 
  column_spec(column = c(1), width = "4cm") %>%
  column_spec(column = c(2:6), width = "3cm") %>%
  footnote(general = "Source: ISSP 2017, WID, WIID and ILO. Variables in original scale",
           footnote_as_chunk = T);tab_an_countries

kableExtra::save_kable(tab_an_countries, here::here("output/tables/tableA04.html"))
```

```{r tableA5, results='asis', cache=TRUE}
set1 <-  RColorBrewer::brewer.pal(n = 8, name = "Set1")
options(ggplot2.discrete.fill = set1)
options(ggplot2.discrete.colour = set1)
ggplot2::theme_set(ggplot2::theme_bw())
ggplot2::theme_update(text=ggplot2::element_text(size=15,  family="serif"))
options(scipen=999)
if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,sjmisc,sjlabelled,haven,vtable)  
rm(list=ls())
tabformat = if(is_html_output()) {"html"} else if(is_latex_output()) {"latex"}
load(here::here("input/data/proc/study1_country.RData"));df1 <- df2

dfreg <- df1 %>% dplyr::select(
  redist,prefineq,
  egal = egal2,
  dclass3res_V,dclass6res,
  homclass3_V_res,
  know_total,
  # Q03pcm,edyears,
  female,
  agenum,
  # partner,
  # workst,
  "gini_disp"=wid_gini_disp,
  abs_red,
  ilo_taxrev,
  ilo_govspe,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  gdppercapita,
  country2,country3,country,
) %>%
  filter(country2 != "SVN") %>% 
  na.omit() %>% 
    mutate(loggdppercapita=gdppercapita/1000
         )

fit_gov<- psych::pca(dfreg[,c("ilo_taxrev","ilo_govspe","abs_red")])
dfreg$wstate<- as.numeric(fit_gov$scores)
dfreg$wstate2 <- ((dfreg$wstate-min(dfreg$wstate,na.rm = T))/(max(dfreg$wstate,na.rm = T)-min(dfreg$wstate,na.rm = T)))*100

dfreg1 <- dfreg %>% dplyr::select(
  egal,
  homclass3_V_res,
  know_total,
  dclass3res_V,
  # Q03pcm,
  # edyears,
  # workst,
  female,
  agenum,
  # partner,
  "gini_disp",
  loggdppercapita,
  wstate2
  ) 

names(dfreg1)
varlab <- 
c("Redistributive preferences","Class-based network homogeneity",
  "Network size","Social class",
  # "Household Income","Education in years","Labor market status",
  "Gender","Age in years",
  # "Has partner",
  "Income Inequality - Gini Index",
  "GDP/capita ",
  "Size of the welfare state")

sjlabelled::set_label(dfreg1) <- varlab

tab05 <- 
st(dfreg1,labels=T,
   title="Table A5: Descriptive statistics for study variables",
   out = "kable",
   numformat = formatfunc(digits = 2),
   digits = 2, 
   factor.counts = F,
   factor.numeric = F,
   summ = list(
     c('notNA(x)','mean(x)','sd(x)','min(x)','max(x)'),
     c('notNA(x)','mean(x)')
   ),
   summ.names = list(
     c('N','Mean','SD','Min','Max'),
     c('Count','Percent')   
     )) %>% 
    kable_classic(html_font = "sans",full_width = F) 

kableExtra::save_kable(tab05, here::here("output/tables/tableA05.html"))
```
