==> processing/04-prod_prep-supplementary-material.Rmd <==
---
title: "Supplementary Analyses"
date: "`r format(Sys.time(), '%A %d %B %Y %X')`"
output: 
  html_document: 
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: no
      number_sections: yes
    code_folding: show  
    number_sections: yes
editor_options: 
  chunk_output_type: console
---

# Setup

```{r setup}
knitr::opts_chunk$set(warning = FALSE, 
                      message = FALSE, 
                      echo = TRUE 
                      )
options(scipen=9999) # desactivar notacion cientifica
if (!require("pacman")) install.packages("pacman") # instalar pacman
pacman::p_load(dplyr,sjmisc,sjPlot,sjlabelled,tidyverse,haven,MLMusingR)
```


# Individual-level

```{r tableS2, results='asis'}
set1 <-  RColorBrewer::brewer.pal(n = 8, name = "Set1")
options(ggplot2.discrete.fill = set1)
options(ggplot2.discrete.colour = set1)
ggplot2::theme_set(ggplot2::theme_bw())
ggplot2::theme_update(text=ggplot2::element_text(size=15,  family="serif"))

if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,sjmisc,ggplot2,interplot,marginaleffects,sjlabelled,haven,stringr,ordinal,texreg,MLMusingR,lme4,gridExtra,grid)  
# rm(list=ls())

load(here::here("input/data/proc/study1_country.RData"));df1 <- df2
dfreg <- df1 %>% dplyr::select(
  prefineq,redist,
  egal = egal2,
  dclass3res_V,class6res,
  n_middle,n_low,n_high,
  homclass3_V_res,
  know_total,
  female,
  agenum,
  WEIGHT,
  region,
  "gini_disp"=wid_gini_disp,
  "gini_mkt",
  gv_spen,
  rel_red,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  rgdpna,
  gdppercapita,
  oecd,
  country2, country, country3,
) %>%
  mutate(logrgdpna = log(rgdpna),
         loggdppercapita=log(gdppercapita)) %>%
  filter(country2 != "SVN")
dfreg <- na.omit(dfreg)
dfreg$egal <- scale(dfreg$egal)

dfreg$n_low <- dfreg$n_low/4 
dfreg$n_middle <- dfreg$n_middle/3
dfreg$n_high <- dfreg$n_high/3

## Model -  EGP3 - DIGCLASS - V to Working -----------------------------$
digclas1_VW_7 <- lmer(egal~1 + dclass3res_V+n_low+female+agenum+ (1|country2),data=dfreg,weights = WEIGHT)
digclas1_VW_8 <- lmer(egal~1 + dclass3res_V+n_low+n_middle+female+agenum+(1|country2),data=dfreg,weights = WEIGHT)
digclas1_VW_10 <- lmer(egal~1 + dclass3res_V+n_low+n_middle+n_high+female+agenum+(1|country2),data=dfreg,weights = WEIGHT)
digclas1_VW_10.1 <- lmer(egal~1 + dclass3res_V*n_low+n_middle+n_high+female+agenum+(1|country2),data=dfreg,weights = WEIGHT)
digclas1_VW_10.2 <- lmer(egal~1 + dclass3res_V*n_middle+n_high+n_low+female+agenum+(1|country2),data=dfreg,weights = WEIGHT)
digclas1_VW_10.3 <- lmer(egal~1 + dclass3res_V*n_high+n_low+n_middle+female+agenum+(1|country2),data=dfreg,weights = WEIGHT)
digclas1_VW_11 <- lmer(egal~1 + dclass3res_V*n_low + dclass3res_V*n_middle + dclass3res_V*n_high +female+agenum+(1|country2),data=dfreg,weights = WEIGHT)


digclas1_VW <- list(digclas1_VW_7,digclas1_VW_8,digclas1_VW_10,
                    digclas1_VW_10.1,digclas1_VW_10.2,digclas1_VW_10.3,digclas1_VW_11)

fit_low<- plot_predictions(digclas1_VW_11, condition = list("dclass3res_V", n_low = seq(0, 1, by = 0.1)),draw=F)
fit_mid<-plot_predictions(digclas1_VW_11, condition = list("dclass3res_V", n_middle = seq(0, 1, by = 0.1)),draw = F)
fit_hig<-plot_predictions(digclas1_VW_11, condition = list("dclass3res_V", n_high = seq(0, 1, by = 0.1)),draw = F)

fit_low$n_low <- as.numeric(as.character(fit_low$n_low))
fit_mid$n_middle <- as.numeric(as.character(fit_mid$n_middle))
fit_hig$n_high <- as.numeric(as.character(fit_hig$n_high))

plot_low <- 
fit_low %>% 
ggplot(aes(y =estimate,x=n_low, fill=dclass3res_V, color=dclass3res_V,ymin=conf.low, ymax=conf.high,group=dclass3res_V)) +
  geom_ribbon(alpha=0.8,size=1,linetype="solid",color="black") +
  geom_line(size=1,color="black") +
  geom_point(size=2,color="black") +
  facet_grid(~factor(dclass3res_V, levels=
                       c("Working Class (V+VI+VII)","Intermediate class (III+IV)","Service Class (I+II)"),
                     c("Working Class","Intermediate Class","Service Class"))) +
  scale_fill_manual(values=c("#323232","#989898","#CCCCCC"))+
  scale_color_manual(values=c("#323232","#989898","#CCCCCC"))+
  scale_x_continuous(
  breaks = c(0, 0.25, 0.5, 0.75, 1),
  labels = c("0", "0.25", "0.50", "0.75", "1"))+
  scale_y_continuous(limits = c(-0.5,0.5))+
  labs(
       x="Ties to Working Class",y="Redistributive preferences",tag = "A") +
  theme(legend.position = "none", legend.title = element_blank())
plot_mid<- 
fit_mid %>% 
ggplot(aes(y =estimate,x=n_middle, fill=dclass3res_V, group=dclass3res_V,ymin=conf.low, ymax=conf.high)) +
  geom_ribbon(alpha=0.8,size=1,linetype="solid",color="black") +
  geom_line(size=1,color="black") +
  geom_point(size=2,color="black") +
  facet_grid(~factor(dclass3res_V, levels=
                       c("Working Class (V+VI+VII)","Intermediate class (III+IV)","Service Class (I+II)"),
                     c("Working Class","Intermediate Class","Service Class"))) +
  scale_fill_manual(values=c("#323232","#989898","#CCCCCC"))+
  scale_color_manual(values=c("#323232","#989898","#CCCCCC"))+
    scale_x_continuous(
  breaks = c(0, 0.25, 0.5, 0.75, 1),
  labels = c("0", "0.25", "0.50", "0.75", "1"))+
  scale_y_continuous(limits = c(-0.5,0.5))+
  labs(
      x="Ties to Intermediate Class",y="Redistributive preferences",tag = "B") +
  theme(legend.position = "none", legend.title = element_blank())

plot_hig <- 
fit_hig %>% 
ggplot(aes(y =estimate,x=n_high, fill=dclass3res_V, group=dclass3res_V,ymin=conf.low, ymax=conf.high)) +
  geom_ribbon(alpha=0.8,size=1,linetype="solid",color="black") +
  geom_line(size=1,color="black") +
  geom_point(size=2,color="black") +
  facet_grid(~factor(dclass3res_V, levels=
                       c("Working Class (V+VI+VII)","Intermediate class (III+IV)","Service Class (I+II)"),
                     c("Working Class","Intermediate Class","Service Class"))) +
  scale_fill_manual(values=c("#323232","#989898","#CCCCCC"))+
  scale_color_manual(values=c("#323232","#989898","#CCCCCC"))+
  scale_x_continuous(
  breaks = c(0, 0.25, 0.5, 0.75, 1),
  labels = c("0", "0.25", "0.50", "0.75", "1"))+
  scale_y_continuous(limits = c(-0.5,0.5))+
  labs(
       x="Ties to Service Class",y="Redistributive preferences",tag = "C") +
  theme(legend.position = "none", legend.title = element_blank())

set.seed(123)

grid.arrange(plot_low,plot_mid,plot_hig, ncol=1)
ggsave(plot = arrangeGrob(plot_low,plot_mid,plot_hig,
             nrow = 3, 
 bottom =textGrob("Source: International Social Survey 2017; the line is the fitted values; CI at 95%",
 gp = gpar(fontface = 1, fontsize = 9),hjust = 1,x = 1
  )),
 filename = "figureS1.png",device = "png",
 path = here::here("output/images"),width = 1,height = 1,units = "cm",scale = 26)


ccoef <- list(homclass3_V_res="Class-based network homogeneity",
              know_total="Network size",
              "dclass3res_VIntermediate class (III+IV)"="Intermediate Class",
              "dclass3res_VWorking Class (V+VI+VII)"="Working Class",
              "n_low" = "Tie to working class",
              "n_middle" = "Tie to Intermediate class",
              "n_high" = "Tie to Service class",
              "dclass3res_VIntermediate class (III+IV):n_low" = "Intermediate × Tie to Working class",
              "dclass3res_VWorking Class (V+VI+VII):n_low"    = "Working × Tie to Working class",
              "dclass3res_VIntermediate class (III+IV):n_middle" = "Intermediate × Tie to Intermediate class",
              "dclass3res_VWorking Class (V+VI+VII):n_middle" = "Working × Tie to Intermediate class",
              "dclass3res_VIntermediate class (III+IV):n_high" = "Intermediate × Tie to Service class",
              "dclass3res_VWorking Class (V+VI+VII):n_high"= "Working × Tie to Service class"
              )

htmlreg(digclas1_VW,stars = c(0.001, 0.01, 0.05, 0.1),single.row = F,symbol = ".",
       file = here::here("output/tables/tableS01.html"),
       caption = paste("Table S1: Multilevel models for class-based profiles, social classs and redistributive preferences"),
       custom.note = "Note: Standard errors in parentheses. %stars.",
       threeparttable = T,
       caption.above = T,leading.zero = T,
       custom.coef.map = ccoef,
      groups = list("Social Class (Ref.= Service Class)" = 1:2,
                    "Network class-profile" = 3:5,
                    "Interactions Class × Class-profile" = 6:11),
      # custom.gof.rows = list("Controls"=c("Yes",rep("Yes",1))),
      include.loglik = FALSE,
      include.aic = FALSE,
      custom.gof.names = c(NA,NA,"Num. groups","Var: Country (Intercept)","Var: Residual")
      )
```

# Country-level controls

```{r data-prep}
set1 <-  RColorBrewer::brewer.pal(n = 8, name = "Set1")
options(ggplot2.discrete.fill = set1)
options(ggplot2.discrete.colour = set1)
ggplot2::theme_set(ggplot2::theme_bw())
ggplot2::theme_update(text=ggplot2::element_text(size=15,  family="serif"))

if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,sjmisc,ggplot2,interplot,marginaleffects,sjlabelled,haven,stringr,ordinal,texreg,MLMusingR,lme4)  
load(here::here("input/data/proc/study1_country.RData"));df1 <- df2
data_corruption <- xlsx::read.xlsx("input/data/original/coruption-data/CPI2017_Full_DataSet-1801.xlsx",sheetIndex = 1)
data_corruption$year <- 2017
data_unionden <- haven::read_dta("input/data/original/union-density/union-density.dta") 
data_migration <- data.table::fread("input/data/original/migration/migrant-stock-share.csv")
data_unemployment <- data.table::fread("input/data/original/unemployment-world-bank/unemployment-rate.csv")
data_qog <- haven::read_dta("input/data/original/QOG/qog_std_ts_jan25.dta") 
data_vdem <- haven::read_dta("input/data/original/QOG/V-Dem-CY-Full+Others-v15.dta") 
data_socx <- data.table::fread("input/data/original/ocde-socx/ocde-socspen.csv")

socx <- 
data_socx %>% 
  filter(SUBJECT =="PUB") %>% 
  dplyr::select(country=LOCATION,year=TIME_PERIOD,socx=OBS_VALUE) %>% 
  mutate(ccode=countrycode::countrycode(country,origin = "iso3c",destination = "iso3n"),
         country=countrycode::countrycode(ccode,origin = "iso3n",destination = "country.name"),
         year=as.numeric(year)) 

union_den <- 
data_unionden %>% 
  dplyr::select(country=ref_area_label,year=time,union_den=obs_value) %>% 
  mutate(ccode=countrycode::countrycode(country,origin = "country.name",destination = "iso3n"),
         country=countrycode::countrycode(ccode,origin = "iso3n",destination = "country.name"),
         year=as.numeric(year)) 

migration<- 
data_migration %>% dplyr::select(country=Entity,year=Year,mig_share=`Share of the population that was born in another country`) %>% 
  mutate(ccode=countrycode::countrycode(country,origin = "country.name",destination = "iso3n"),
         country=countrycode::countrycode(ccode,origin = "iso3n",destination = "country.name"),
         year=as.numeric(year)) 

corruption <- 
data_corruption %>% 
  dplyr::select(country=Country,year,iso3=ISO3,cpi=CPI.Score.2017) %>% 
  mutate(ccode=countrycode::countrycode(country,origin = "country.name",destination = "iso3n"),
         country=countrycode::countrycode(ccode,origin = "iso3n",destination = "country.name"),
         year=as.numeric(year)) 

unemployment <- 
data_unemployment %>% 
  dplyr::select(country=Entity,year=Year,unemployment_rate=`Unemployment, total (% of total labor force) (modeled ILO estimate)`) %>% 
  mutate(ccode=countrycode::countrycode(country,origin = "country.name",destination = "iso3n"),
         country=countrycode::countrycode(ccode,origin = "iso3n",destination = "country.name"),
         year=as.numeric(year)) 

democracy <- 
data_qog %>% 
  dplyr::select(country=cname,year=year,van_index,p_polity2,fe_etfra) %>% 
  mutate(ccode=countrycode::countrycode(country,origin = "country.name",destination = "iso3n"),
         country=countrycode::countrycode(ccode,origin = "iso3n",destination = "country.name"),
         year=as.numeric(year)) 

vdem <- 
data_vdem %>% 
  dplyr::select(country=country_name,year=year,v2x_delibdem,v2x_polyarchy,v2x_egal,v2catrauni) %>% 
  mutate(ccode=countrycode::countrycode(country,origin = "country.name",destination = "iso3n"),
         country=countrycode::countrycode(ccode,origin = "iso3n",destination = "country.name"),
         year=as.numeric(year)) 

additional_macro <- 
union_den %>% 
  full_join(migration,by=c("ccode","country","year")) %>% 
  full_join(corruption,by=c("ccode","country","year")) %>% 
  full_join(unemployment,by=c("ccode","country","year")) %>%   
  full_join(democracy,by=c("ccode","country","year")) %>%
  full_join(vdem,by=c("ccode","country","year")) %>%  
  full_join(socx,by=c("ccode","country","year")) %>% 
  arrange(year)

summary(additional_macro)

country_code<- as.numeric(levels(as.factor(as.character(df1$ccode))))

additional_macro_2000to2018 <- 
  additional_macro %>% 
  filter(year %in% c(2015:2017)) %>% 
  filter(ccode %in% country_code)


additional_macro_avg <- 
additional_macro_2000to2018 %>% 
  group_by(ccode) %>% 
  summarise(union_den=mean(union_den,na.rm=T),
            mig_share=mean(mig_share,na.rm=T),
            cpi=mean(cpi,na.rm=T),
            unemployment_rate=mean(unemployment_rate,na.rm=T),
            van_index=mean(van_index,na.rm=T),
            p_polity2=mean(p_polity2,na.rm=T),            
            v2x_delibdem=mean(v2x_delibdem ,na.rm=T),
            v2x_polyarchy=mean(v2x_polyarchy ,na.rm=T),
            v2x_egal = mean(v2x_egal,na.rm=T),
            fe_etfra = mean(fe_etfra,na.rm=T),
            v2catrauni = mean(v2catrauni,na.rm=T),
            socx = mean(socx,na.rm=T)
            )
summary(additional_macro_avg)
save(additional_macro_avg,file ="input/data/proc/additional_macro_avg.RData")
```

```{r tableS2}
set1 <-  RColorBrewer::brewer.pal(n = 8, name = "Set1")
options(ggplot2.discrete.fill = set1)
options(ggplot2.discrete.colour = set1)
ggplot2::theme_set(ggplot2::theme_bw())
ggplot2::theme_update(text=ggplot2::element_text(size=15,  family="serif"))
options(scipen=999)
if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,sjmisc,ggplot2,interplot,marginaleffects,sjlabelled,haven,stringr,ordinal,texreg,MLMusingR,lme4)  
# rm(list=ls()) 

load(here::here("input/data/proc/study1_country.RData"));df1 <- df2
load(here::here("input/data/proc/additional_macro_avg.RData"))
df1$idnum <- rownames(df1)
dfid <- df1 %>% dplyr::select(
  idnum,
  redist,prefineq,
  egal = egal2,
  dclass3res_V,dclass6res,
  homclass3_V_res,
  know_total,
  # Q03pcm,
  # edyears,
  female,
  agenum,
  # partner,
  # workst,
  WEIGHT,
  region,
  "gini_disp"=wid_gini_disp,
  "gini_mkt",
  gv_spen,
  rel_red,
  abs_red,
  ilo_taxrev,
  ilo_govspe,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  rgdpna,
  gdppercapita,
  country2, country, country3,
) %>% na.omit()

# Merge new macro-data
df1 <- df1 %>% left_join(.,additional_macro_avg,by ="ccode")
dfreg <- df1 %>% dplyr::select(
  idnum,
  redist,prefineq,
  egal = egal2,
  dclass3res_V,
  homclass3_V_res,dclass6res,
  know_total,
  # Q03pcm,
  # edyears,
  female,
  agenum,
  # partner,
  # workst,
  WEIGHT,
  region,
  "gini_disp"=wid_gini_disp,
  "gini_mkt",
  gv_spen,
  rel_red,
  abs_red,
  ilo_taxrev,
  ilo_govspe,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  rgdpna,
  gdppercapita,
  cpi,
  mig_share,
  union_den,
  v2x_polyarchy,
  unemployment_rate,
  v2catrauni,socx,
  country2, country, country3,
) %>%
  filter(country2 != "SVN") %>% 
  # na.omit() %>%
    mutate(loggdppercapita=scale(gdppercapita/1000),
         age2 = agenum ^ 2) %>% 
  filter(idnum %in%  dfid$idnum) 

fit_gov<- psych::pca(dfreg[,c("ilo_taxrev","ilo_govspe","abs_red")])
dfreg$wstate<- as.numeric(fit_gov$scores)
dfreg$wstate2 <- ((dfreg$wstate-min(dfreg$wstate,na.rm = T))/(max(dfreg$wstate,na.rm = T)-min(dfreg$wstate,na.rm = T)))*100
dfreg$wstate2 <- as.numeric(scale(dfreg$wstate2))

dfreg$socx<- as.numeric(scale(dfreg$socx))
dfreg$gini_disp <- as.numeric(scale(dfreg$gini_disp))
dfreg$ilo_taxrev <- as.numeric(scale(dfreg$ilo_taxrev))
dfreg$ilo_govspe <- as.numeric(scale(dfreg$ilo_govspe))
dfreg$cpi <- as.numeric(scale(dfreg$cpi))
dfreg$mig_share <- as.numeric(scale(dfreg$mig_share))
dfreg$union_den <- as.numeric(scale(dfreg$union_den))
dfreg$v2x_polyarchy <- as.numeric(scale(dfreg$v2x_polyarchy))
dfreg$unemployment_rate <- as.numeric(scale(dfreg$unemployment_rate))
dfreg$egal <- scale(dfreg$egal)

# Variable Group-Centering ------------------------------------------------$
dfreg <- 
  dfreg %>% 
  mutate(to_dummy(female)
         )
dfreg$female_gc = group_center(dfreg$female_2, grp = dfreg$country2)
dfreg$agenum_gc = group_center(dfreg$agenum, grp = dfreg$country2)
dfreg$homclass3_V_gc <-group_center(dfreg$homclass3_V_res, grp = dfreg$country2)
dfreg$know_total_gc = group_center(dfreg$know_total, grp = dfreg$country2)

## Model -  EGP3 - DIGCLASS - V to Working -----------------------------$
homo_V_gini_gdp_wstate2 <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         gini_disp +loggdppercapita + wstate2+
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

homo_V_gini_gdp_cpi <-
  lmer(egal~
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         gini_disp +loggdppercapita + wstate2 + cpi +
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

homo_V_gini_gdp_mig <-
  lmer(egal~
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         gini_disp +loggdppercapita + wstate2 + mig_share +
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

homo_V_gini_gdp_union <-
  lmer(egal~
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         gini_disp +loggdppercapita + wstate2 +union_den + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

homo_V_gini_gdp_unempl <-
  lmer(egal~
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         gini_disp +loggdppercapita + wstate2 +unemployment_rate + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

homo_V_gini_gdp_v2x_polyarchy <-
  lmer(egal~
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         gini_disp +loggdppercapita + wstate2 + v2x_polyarchy +
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

homo_V_gini_full2 <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         gini_disp +loggdppercapita + wstate2+
         homclass3_V_gc*
         dclass3res_V*
         gini_disp + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

homo_V_gini_full3 <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         gini_disp +loggdppercapita + wstate2+
         homclass3_V_gc*
         dclass3res_V*
         gini_disp + mig_share + union_den +unemployment_rate + v2x_polyarchy + cpi+ 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

models_macro <- list(homo_V_gini_gdp_wstate2,
                     homo_V_gini_gdp_cpi,
                     homo_V_gini_gdp_v2x_polyarchy,                     
                     homo_V_gini_gdp_mig,
                     homo_V_gini_gdp_union,
                     homo_V_gini_gdp_unempl,
                     homo_V_gini_full2,
                     homo_V_gini_full3
                     )

screenreg(models_macro)

ccoef <- list(
              "dclass3res_VIntermediate class (III+IV)"="Intermediate Class",
              "dclass3res_VWorking Class (V+VI+VII)"="Working Class",
              homclass3_V_gc="Class-based network homogeneity (CWC)",
              "gini_disp" = "Income inequality (Gini index)",
              loggdppercapita = "GDP/capita",
              wstate2="Size of the welfare state",
              cpi = "Corruption Index",
              v2x_polyarchy = "Democracy (polyarchy index)",
              mig_share = "Migration (pop %)",
              union_den = "Union density",
              unemployment_rate = "Unemployment rate",
              "homclass3_V_gc:dclass3res_VIntermediate class (III+IV)" = "Intermediate Class × Homogeneity",
              "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII)" = "Working Class × Homogeneity",
              "homclass3_V_gc:gini_disp"= "Homogeneity x Income Inequality",
              "dclass3res_VIntermediate class (III+IV):gini_disp" = "Intermediate Class × Income Inequality",
              "dclass3res_VWorking Class (V+VI+VII):gini_disp" = "Working Class × Income Inequality",
              "homclass3_V_gc:dclass3res_VIntermediate class (III+IV):gini_disp" = "Intermediate Class × Homogeneity × Income Inequality",
              "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII):gini_disp" = "Working Class × Homogeneity × Income Inequality")

htmlreg(models_macro,stars = c(0.001, 0.01, 0.05, 0.1),scalebox = 0.7,symbol = ".",
       file = here::here("output/tables/tableS02.html"),
       caption = paste("Table S2: Multilevel models for country-level factors,
                       income inequality, network homogeneity and redistributive preferences"),
       custom.note = " Note: All individual level controls centered within cluster (group mean).  Standard errors in parentheses. %stars. The depedent variable and country-level variables are in standardized (z-score)",
       caption.above = T,leading.zero = T,single.row = T,threeparttable = T,
       custom.coef.map = ccoef,
      # groups = list("Social Class (Ref.= Service Class)" = 2:3,"Homogeneity x Social Class"=7:8),
      # custom.gof.rows = list("Controls"=rep("Yes",4)),
      include.loglik = FALSE,include.aic = FALSE,
      custom.gof.names = c(NA,NA,"Num. groups",
                           "Var: Country (Intercept)",
                           "Var: Country Homogeneity",
                           "Var: Country Intermediate Class ",
                           "Var: Country Working Class",
                           "Cov: Country (Intercept), Homogeneity",
                           "Cov: Country (Intercept), Intermediate Class ",
                           "Cov: Country (Intercept), Working Class",
                           "Cov: Country Homogeneity, Intermediate Class",
                           "Cov: Country Homogeneity, Working Class",
                           "Cov: Country Intermediate Class, Working Class",
                           "Var: Residual")
      )
```

# Cross-level interaction 

```{r tableS3}
# Main cross-level interaction for Income Inequality---------------------------
int_xlvl_gini <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         homclass3_V_gc*
         dclass3res_V*
         gini_disp + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)
# Main cross-level interaction with Welfare State Size--------------------------
int_xlvl_wstate <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         gini_disp +
         homclass3_V_gc*
         dclass3res_V*
         wstate2 + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

# Double cross-level interaction Gini and Welfare State Size (Tax Reveneu)-------
int_xlvl_gini_taxrev <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         homclass3_V_gc*dclass3res_V*gini_disp + 
         homclass3_V_gc*dclass3res_V*ilo_taxrev + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

# Double cross-level interaction Gini and Welfare State Size (Goverment Spending)------
int_xlvl_gini_govspe <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         homclass3_V_gc*dclass3res_V*gini_disp + 
         homclass3_V_gc*dclass3res_V*ilo_govspe + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)
# Double cross-level interaction Gini and Welfare State Size (Redistribution)----
int_xlvl_gini_redist <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         homclass3_V_gc*dclass3res_V*gini_disp + 
         homclass3_V_gc*dclass3res_V*abs_red + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

# Double cross-level interaction Gini and Welfare State Size (Index)------------
int_xlvl_gini_wstate <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         homclass3_V_gc*dclass3res_V*gini_disp + 
         homclass3_V_gc*dclass3res_V*wstate2 + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)


# Double cross-level interaction Gini and GDP ----------------------------------
int_xlvl_gini_gdp<- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         homclass3_V_gc*dclass3res_V*gini_disp + 
         homclass3_V_gc*dclass3res_V*loggdppercapita +
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)


models_interactions<- list(int_xlvl_gini,int_xlvl_gini_taxrev,int_xlvl_gini_govspe,int_xlvl_gini_redist,int_xlvl_gini_wstate,int_xlvl_gini_gdp)

ccoef <- list(
  # Class
  "dclass3res_VIntermediate class (III+IV)" = "Intermediate Class",
  "dclass3res_VWorking Class (V+VI+VII)" = "Working Class",
  
  # Homogeneity
  homclass3_V_gc = "Class-Based Network Homogeneity (CWC)",
  
  # Class × Homogeneity
  "homclass3_V_gc:dclass3res_VIntermediate class (III+IV)" = "Intermediate Class × Homogeneity",
  "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII)" = "Working Class × Homogeneity",

  # Macro variables
  gini_disp = "Income Inequality (Gini Index)",
  ilo_taxrev = "Tax Revenue (% of GDP)",
  ilo_govspe = "Government Spending (% of GDP)",
  abs_red = "Redistribution",
  wstate2 = "Welfare State Size",
  loggdppercapita = "GDP per Capita (Log)",
  
  # Homogeneity × Macro
  "homclass3_V_gc:gini_disp" = "Homogeneity × Income Inequality",
  "dclass3res_VIntermediate class (III+IV):gini_disp" = "Intermediate Class × Income Inequality",
  "dclass3res_VWorking Class (V+VI+VII):gini_disp" = "Working Class × Income Inequality",
  "homclass3_V_gc:dclass3res_VIntermediate class (III+IV):gini_disp" = "Intermediate Class × Homogeneity × Income Inequality",
  "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII):gini_disp" = "Working Class × Homogeneity × Income Inequality",
  
  "homclass3_V_gc:ilo_taxrev" = "Homogeneity × Tax Revenue",
  "dclass3res_VIntermediate class (III+IV):ilo_taxrev" = "Intermediate Class × Tax Revenue",
  "dclass3res_VWorking Class (V+VI+VII):ilo_taxrev" = "Working Class × Tax Revenue",  
  "homclass3_V_gc:dclass3res_VIntermediate class (III+IV):ilo_taxrev" = "Intermediate Class × Homogeneity × Tax Revenue",
  "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII):ilo_taxrev" = "Working Class × Homogeneity × Tax Revenue",
  
  "homclass3_V_gc:ilo_govspe" = "Homogeneity × Government Spending",
  "dclass3res_VIntermediate class (III+IV):ilo_govspe" = "Intermediate Class × Government Spending",
  "dclass3res_VWorking Class (V+VI+VII):ilo_govspe" = "Working Class × Government Spending",  
  "homclass3_V_gc:dclass3res_VIntermediate class (III+IV):ilo_govspe" = "Intermediate Class × Homogeneity × Government Spending",
  "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII):ilo_govspe" = "Working Class × Homogeneity × Government Spending",
  
  "homclass3_V_gc:abs_red" = "Homogeneity × Redistribution",
  "dclass3res_VIntermediate class (III+IV):abs_red" = "Intermediate Class × Redistribution",
  "dclass3res_VWorking Class (V+VI+VII):abs_red" = "Working Class × Redistribution",    
  "homclass3_V_gc:dclass3res_VIntermediate class (III+IV):abs_red" = "Intermediate Class × Homogeneity × Redistribution",
  "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII):abs_red" = "Working Class × Homogeneity × Redistribution",
  
  "homclass3_V_gc:wstate2" = "Homogeneity × Welfare State Size",
  "dclass3res_VIntermediate class (III+IV):wstate2" = "Intermediate Class × Welfare State Size",
  "dclass3res_VWorking Class (V+VI+VII):wstate2" = "Working Class × Welfare State Size",    
  "homclass3_V_gc:dclass3res_VIntermediate class (III+IV):wstate2" = "Intermediate Class × Homogeneity × Welfare State Size",
  "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII):wstate2" = "Working Class × Homogeneity × Welfare State Size",
  
  "homclass3_V_gc:loggdppercapita" = "Homogeneity × GDP per Capita",
  "dclass3res_VIntermediate class (III+IV):loggdppercapita" = "Intermediate Class × GDP per Capita",
  "dclass3res_VWorking Class (V+VI+VII):loggdppercapita" = "Working Class × GDP per Capita",      
  "homclass3_V_gc:dclass3res_VIntermediate class (III+IV):loggdppercapita" = "Intermediate Class × Homogeneity × GDP per Capita",
  "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII):loggdppercapita" = "Working Class × Homogeneity × GDP per Capita"
)

screenreg(models_interactions)
htmlreg(models_interactions,stars = c(0.001, 0.01, 0.05, 0.1),scalebox = 0.7,symbol = ".",
       file = here::here("output/tables/tableS03.html"),
       caption = paste("Table S3: Robustness analysis of the multilevel models for income inequality, network homogeneity and redistributive preferences"),
       custom.note = " Note: All individual level controls centered within cluster (group mean).  Standard errors in parentheses. %stars",
       caption.above = T,leading.zero = T,single.row = T,threeparttable = T,
       custom.coef.map = ccoef,
      # groups = list("Social Class (Ref.= Service Class)" = 2:3,"Homogeneity x Social Class"=7:8),
      # custom.gof.rows = list("Controls"=rep("Yes",4)),
      include.loglik = FALSE,include.aic = FALSE,
      custom.gof.names = c(NA,NA,"Num. groups",
                           "Var: Country (Intercept)",
                           "Var: Country Homogeneity",
                           "Var: Country Intermediate Class ",
                           "Var: Country Working Class",
                           "Cov: Country (Intercept), Homogeneity",
                           "Cov: Country (Intercept), Intermediate Class ",
                           "Cov: Country (Intercept), Working Class",
                           "Cov: Country Homogeneity, Intermediate Class",
                           "Cov: Country Homogeneity, Working Class",
                           "Cov: Country Intermediate Class, Working Class",
                           "Var: Residual")
      )
```

```{r tableS4, results='asis'}
set1 <-  RColorBrewer::brewer.pal(n = 8, name = "Set1")
options(ggplot2.discrete.fill = set1)
options(ggplot2.discrete.colour = set1)
ggplot2::theme_set(ggplot2::theme_bw())
ggplot2::theme_update(text=ggplot2::element_text(size=15,  family="serif"))
if (!require("pacman")) install.packages("pacman")
pacman::p_load(dplyr,sjmisc,ggplot2,interplot,marginaleffects,sjlabelled,haven,stringr,ordinal,texreg,MLMusingR,lme4)  
rm(list=ls())
load(here::here("input/data/proc/study1_country.RData"));df1 <- df2
dfreg <- df1 %>% dplyr::select(
  prefineq,redist,
  egal = egal2,
  dclass3res_V,
  homclass3_V_res,dclass6res,
  know_total,
  female,
  agenum,
  WEIGHT,
  region,
  "gini_disp"=wid_gini_disp,
  "gini_mkt",
  ilo_taxrev,ilo_govspe,abs_red,
  rel_red,
  d10d1=wid_p90p10,
  wid_p90p50,
  top10=wid_sharetop10,
  rgdpna,
  gdppercapita,
  oecd,
  country2, country, country3,
) %>%
  mutate(logrgdpna = log(rgdpna),
         loggdppercapita=log(gdppercapita)
         ) %>%
  filter(country2 != "SVN") 
dfreg$egal <- as.numeric(scale(dfreg$egal))
dfreg <- na.omit(dfreg)

fit_gov<- psych::pca(dfreg[,c("ilo_taxrev","ilo_govspe","abs_red")])
dfreg$wstate<- as.numeric(fit_gov$scores)
dfreg$wstate2 <- ((dfreg$wstate-min(dfreg$wstate,na.rm = T))/(max(dfreg$wstate,na.rm = T)-min(dfreg$wstate,na.rm = T)))*100

## Model -  EGP3 - DIGCLASS - V to Working -----------------------------$ 
digclas1_VW_0<- lmer(egal~1 + dclass6res + female+agenum+(1|country2),data=dfreg,weights = WEIGHT)
digclas1_VW_1<- lmer(egal~1 + homclass3_V_res+know_total+(1|country2),data=dfreg,weights = WEIGHT)
digclas1_VW_3 <- update(digclas1_VW_1, . ~ . + dclass6res)
digclas1_VW_7 <- update(digclas1_VW_3, . ~ . +dclass6res*homclass3_V_res)
digclas1_VW_8 <- update(digclas1_VW_3, . ~ . +dclass6res*homclass3_V_res*gini_disp)
digclas1_VW <- list(digclas1_VW_0,digclas1_VW_1,
                    digclas1_VW_3,
                    # digclas1_VW_4,
                    #digclas1_VW_5,
                    # digclas1_VW_6,
                    digclas1_VW_7)

ccoef <- list("dclass6resLower Service class"="Lower Service",
              "dclass6resRoutine nonmanual class" = "Routine nonmanual",                 
              "dclass6resSelf-employed" = "Self-employed",                           
              "dclass6resSkilled working class" = "Skilled working",
              "dclass6resUnskilled working class" = "Unskilled working",
              homclass3_V_res="Class-based network homogeneity",
              know_total="Network size",              
              "homclass3_V_res:dclass6resLower Service class" = "Lower Service × Homogeneity",
              "homclass3_V_res:dclass6resRoutine nonmanual class" = "Routine nonmanual × Homogeneity",
              "homclass3_V_res:dclass6resSelf-employed"= "Self-employed × Homogeneity",
              "homclass3_V_res:dclass6resSkilled working class" = "Skilled working class × Homogeneity",
              "homclass3_V_res:dclass6resUnskilled working class" = "Unskilled working class × Homogeneity" 
              )

screenreg(digclas1_VW)

htmlreg(digclas1_VW,stars = c(0.001, 0.01, 0.05, 0.1),single.row = T,symbol = ".",
       caption = paste("Table S4: Multilevel models for network homogeneity and redistributive preferences (6-class version)"),
       custom.note = "Note: Standard errors in parentheses. %stars",
       threeparttable = T,
       caption.above = T,leading.zero = T,
       custom.coef.map = ccoef,
       groups =
         list("Social Class (Ref.= Upper Service)" = 1:5,
              "Social Class × Homogeneity"=8:12),
      custom.gof.rows = list("Controls"=c("Yes",rep("Yes",3))),include.loglik = FALSE,include.aic = FALSE,
      custom.gof.names = c(NA,NA,"Num. groups","Var: Country (Intercept)","Var: Residual"),
      file = here::here("output/tables/tableS04.html")
      )
```

```{r tableS5}
# Main cross-level interaction for Income Inequality---------------------------
int_xlvl_gini <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass6res+
         female_gc+agenum_gc+
         homclass3_V_gc*
         dclass6res*
         gini_disp + 
         (homclass3_V_gc+dclass6res|country2),data=dfreg,weights = WEIGHT)
# Main cross-level interaction with Welfare State Size--------------------------
int_xlvl_wstate <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass6res+
         female_gc+agenum_gc+
         gini_disp +
         homclass3_V_gc*
         dclass6res*
         wstate2 + 
         (homclass3_V_gc+dclass6res|country2),data=dfreg,weights = WEIGHT)

# Double cross-level interaction Gini and Welfare State Size (Tax Reveneu)-------
int_xlvl_gini_taxrev <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass6res+
         female_gc+agenum_gc+
         homclass3_V_gc*dclass6res*gini_disp + 
         homclass3_V_gc*dclass6res*ilo_taxrev + 
         (homclass3_V_gc+dclass6res|country2),data=dfreg,weights = WEIGHT)

# Double cross-level interaction Gini and Welfare State Size (Goverment Spending)------
int_xlvl_gini_govspe <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass6res+
         female_gc+agenum_gc+
         homclass3_V_gc*dclass6res*gini_disp + 
         homclass3_V_gc*dclass6res*ilo_govspe + 
         (homclass3_V_gc+dclass6res|country2),data=dfreg,weights = WEIGHT)
# Double cross-level interaction Gini and Welfare State Size (Redistribution)----
int_xlvl_gini_redist <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass6res+
         female_gc+agenum_gc+
         homclass3_V_gc*dclass6res*gini_disp + 
         homclass3_V_gc*dclass6res*abs_red + 
         (homclass3_V_gc+dclass6res|country2),data=dfreg,weights = WEIGHT)

# Double cross-level interaction Gini and Welfare State Size (Index)------------
int_xlvl_gini_wstate <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass6res+
         female_gc+agenum_gc+
         homclass3_V_gc*dclass6res*gini_disp + 
         homclass3_V_gc*dclass6res*wstate2 + 
         (homclass3_V_gc+dclass6res|country2),data=dfreg,weights = WEIGHT)


# Double cross-level interaction Gini and GDP ----------------------------------
int_xlvl_gini_gdp<- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass6res+
         female_gc+agenum_gc+
         homclass3_V_gc*dclass6res*gini_disp + 
         homclass3_V_gc*dclass6res*loggdppercapita +
         (homclass3_V_gc+dclass6res|country2),data=dfreg,weights = WEIGHT)


models_interactions<- list(int_xlvl_gini,int_xlvl_gini_taxrev,int_xlvl_gini_govspe,int_xlvl_gini_redist,int_xlvl_gini_wstate,int_xlvl_gini_gdp)

ccoef <- list(homclass3_V_gc="Class-based network homogeneity",
              know_total="Network size",
              "dclass6resLower Service class"="Lower Service",
              "dclass6resRoutine nonmanual class" = "Routine nonmanual",                 
              "dclass6resSelf-employed" = "Self-employed",                           
              "dclass6resSkilled working class" = "Skilled working",
              "dclass6resUnskilled working class" = "Unskilled working",
              "homclass3_V_gc:dclass6resLower Service class" = "Homogeneity × Lower Service",
              "homclass3_V_gc:dclass6resRoutine nonmanual class" = "Homogeneity × Routine nonmanual",
              "homclass3_V_gc:dclass6resSelf-employed"= "Homogeneity × Self-employed",
              "homclass3_V_gc:dclass6resSkilled working class" = "Homogeneity × Skilled working class",
              "homclass3_V_gc:dclass6resUnskilled working class" = "Homogeneity × Unskilled working class",
              
              gini_disp = "Income inequality (Gini index)",
              ilo_taxrev="(i) Tax revenue",
              ilo_govspe="(ii) Gov. Spending",
              abs_red="(iii) Redistribution",              
              wstate2="(i,ii,iii) Size of the welfare state",            
              loggdppercapita = "GDP/capita",
              
              # 3-way interactions: × Gini
              "homclass3_V_gc:gini_disp" = "Homogeneity × Gini",
              "dclass6resLower Service class:gini_disp"     = "Lower Service × Gini",
              "dclass6resRoutine nonmanual class:gini_disp" = "Routine Nonmanual × Gini",
              "dclass6resSelf-employed:gini_disp"           = "Self-employed × Gini",
              "dclass6resSkilled working class:gini_disp"   = "Skilled Working Class × Gini",
              "dclass6resUnskilled working class:gini_disp" = "Unskilled Working Class × Gini",
              "homclass3_V_gc:dclass6resLower Service class:gini_disp"         = "Homogeneity × Lower Service × Gini",
              "homclass3_V_gc:dclass6resRoutine nonmanual class:gini_disp"     = "Homogeneity × Routine Nonmanual × Gini",
              "homclass3_V_gc:dclass6resSelf-employed:gini_disp"               = "Homogeneity × Self-employed × Gini",
              "homclass3_V_gc:dclass6resSkilled working class:gini_disp"       = "Homogeneity × Skilled Working Class × Gini",
              "homclass3_V_gc:dclass6resUnskilled working class:gini_disp"     = "Homogeneity × Unskilled Working Class × Gini",
  
              # 3-way interactions: × Tax Revenue
              "homclass3_V_gc:ilo_taxrev" = "Homogeneity × Tax Revenue",
              "dclass6resLower Service class:ilo_taxrev"     = "Lower Service × Tax Revenue",
              "dclass6resRoutine nonmanual class:ilo_taxrev" = "Routine Nonmanual × Tax Revenue",
              "dclass6resSelf-employed:ilo_taxrev"           = "Self-employed × Tax Revenue",
              "dclass6resSkilled working class:ilo_taxrev"   = "Skilled Working Class × Tax Revenue",
              "dclass6resUnskilled working class:ilo_taxrev" = "Unskilled Working Class × Tax Revenue",              
              "homclass3_V_gc:dclass6resLower Service class:ilo_taxrev"         = "Homogeneity × Lower Service × Tax Revenue",
              "homclass3_V_gc:dclass6resRoutine nonmanual class:ilo_taxrev"     = "Homogeneity × Routine Nonmanual × Tax Revenue",
              "homclass3_V_gc:dclass6resSelf-employed:ilo_taxrev"               = "Homogeneity × Self-employed × Tax Revenue",
              "homclass3_V_gc:dclass6resSkilled working class:ilo_taxrev"       = "Homogeneity × Skilled Working Class × Tax Revenue",
              "homclass3_V_gc:dclass6resUnskilled working class:ilo_taxrev"     = "Homogeneity × Unskilled Working Class × Tax Revenue",
              
              # 3-way interactions: × Government Spending
              "homclass3_V_gc:ilo_govspe" = "Homogeneity × Gov. Spending",
              "dclass6resLower Service class:ilo_govspe"     = "Lower Service × Gov. Spending",
              "dclass6resRoutine nonmanual class:ilo_govspe" = "Routine Nonmanual × Gov. Spending",
              "dclass6resSelf-employed:ilo_govspe"           = "Self-employed × Gov. Spending",
              "dclass6resSkilled working class:ilo_govspe"   = "Skilled Working Class × Gov. Spending",
              "dclass6resUnskilled working class:ilo_govspe" = "Unskilled Working Class × Gov. Spending",                
              "homclass3_V_gc:dclass6resLower Service class:ilo_govspe"         = "Homogeneity × Lower Service × Gov. Spending",
              "homclass3_V_gc:dclass6resRoutine nonmanual class:ilo_govspe"     = "Homogeneity × Routine Nonmanual × Gov. Spending",
              "homclass3_V_gc:dclass6resSelf-employed:ilo_govspe"               = "Homogeneity × Self-employed × Gov. Spending",
              "homclass3_V_gc:dclass6resSkilled working class:ilo_govspe"       = "Homogeneity × Skilled Working Class × Gov. Spending",
              "homclass3_V_gc:dclass6resUnskilled working class:ilo_govspe"     = "Homogeneity × Unskilled Working Class × Gov. Spending",
  
              # 3-way interactions: × Redistribution
              "homclass3_V_gc:abs_red" = "Homogeneity × Redistribution",
              "dclass6resLower Service class:abs_red"     = "Lower Service × Redistribution",
              "dclass6resRoutine nonmanual class:abs_red" = "Routine Nonmanual × Redistribution",
              "dclass6resSelf-employed:abs_red"           = "Self-employed × Redistribution",
              "dclass6resSkilled working class:abs_red"   = "Skilled Working Class × Redistribution",
              "dclass6resUnskilled working class:abs_red" = "Unskilled Working Class × Redistribution",                
              "homclass3_V_gc:dclass6resLower Service class:abs_red"         = "Homogeneity × Lower Service × Redistribution",
              "homclass3_V_gc:dclass6resRoutine nonmanual class:abs_red"     = "Homogeneity × Routine Nonmanual × Redistribution",
              "homclass3_V_gc:dclass6resSelf-employed:abs_red"               = "Homogeneity × Self-employed × Redistribution",
              "homclass3_V_gc:dclass6resSkilled working class:abs_red"       = "Homogeneity × Skilled Working Class × Redistribution",
              "homclass3_V_gc:dclass6resUnskilled working class:abs_red"     = "Homogeneity × Unskilled Working Class × Redistribution",
  
              # 3-way interactions: × Welfare State
              "homclass3_V_gc:wstate2" = "Homogeneity × Welfare State",
              "dclass6resLower Service class:wstate2"     = "Lower Service × Welfare State",
              "dclass6resRoutine nonmanual class:wstate2" = "Routine Nonmanual × Welfare State",
              "dclass6resSelf-employed:wstate2"           = "Self-employed × Welfare State",
              "dclass6resSkilled working class:wstate2"   = "Skilled Working Class × Welfare State",
              "dclass6resUnskilled working class:wstate2" = "Unskilled Working Class × Welfare State",                   
              "homclass3_V_gc:dclass6resLower Service class:wstate2"         = "Homogeneity × Lower Service × Welfare State",
              "homclass3_V_gc:dclass6resRoutine nonmanual class:wstate2"     = "Homogeneity × Routine Nonmanual × Welfare State",
              "homclass3_V_gc:dclass6resSelf-employed:wstate2"               = "Homogeneity × Self-employed × Welfare State",
              "homclass3_V_gc:dclass6resSkilled working class:wstate2"       = "Homogeneity × Skilled Working Class × Welfare State",
              "homclass3_V_gc:dclass6resUnskilled working class:wstate2"     = "Homogeneity × Unskilled Working Class × Welfare State",
              # 3-way interactions: × GDP
              "homclass3_V_gc:loggdppercapita" = "Homogeneity × GDP",
              "dclass6resLower Service class:loggdppercapita"     = "Lower Service × GDP",
              "dclass6resRoutine nonmanual class:loggdppercapita" = "Routine Nonmanual × GDP",
              "dclass6resSelf-employed:loggdppercapita"           = "Self-employed × GDP",
              "dclass6resSkilled working class:loggdppercapita"   = "Skilled Working Class × GDP",
              "dclass6resUnskilled working class:loggdppercapita" = "Unskilled Working Class × GDP",               
              "homclass3_V_gc:dclass6resLower Service class:loggdppercapita"         = "Homogeneity × Lower Service × GDP",
              "homclass3_V_gc:dclass6resRoutine nonmanual class:loggdppercapita"     = "Homogeneity × Routine Nonmanual × GDP",
              "homclass3_V_gc:dclass6resSelf-employed:loggdppercapita"               = "Homogeneity × Self-employed × GDP",
              "homclass3_V_gc:dclass6resSkilled working class:loggdppercapita"       = "Homogeneity × Skilled Working Class × GDP",
              "homclass3_V_gc:dclass6resUnskilled working class:loggdppercapita"     = "Homogeneity × Unskilled Working Class × GDP"              
              )

htmlreg(models_interactions,stars = c(0.001, 0.01, 0.05, 0.1),scalebox = 0.7,symbol = ".",
       file = here::here("output/tables/tableS05.html"),
       caption = paste("Table S5: Robustness analysis of the multilevel models for income inequality, network homogeneity and redistributive preferences"),
       custom.note = " Note: All individual level controls centered within cluster (group mean).  Standard errors in parentheses. %stars",
       caption.above = T,leading.zero = T,single.row = T,threeparttable = T,
       custom.coef.map = ccoef,
      # groups = list("Social Class (Ref.= Service Class)" = 2:3,"Homogeneity x Social Class"=7:8),
      # custom.gof.rows = list("Controls"=rep("Yes",4)),
      include.loglik = FALSE,include.aic = FALSE,
      custom.gof.names = c(NA,NA,
  "Num. groups: country2" = "Number of countries",
  # Variances
  "Var: country2 (Intercept)" = "Var: country Intercept",
  "Var: country2 homclass3_V_gc" = "Var: country Homogeneity",
  "Var: country2 dclass6resLower Service class" = "Var: country Lower Service",
  "Var: country2 dclass6resRoutine nonmanual class" = "Var: country Routine Nonmanual",
  "Var: country2 dclass6resSelf-employed" = "Var: country Self-employed",
  "Var: country2 dclass6resSkilled working class" = "Var: country Skilled Working Class",
  "Var: country2 dclass6resUnskilled working class" = "Var: country Unskilled Working Class",

  # Covariances with intercept
  "Cov: country2 (Intercept) homclass3_V_gc" = "Cov: Intercept × Homogeneity",
  "Cov: country2 (Intercept) dclass6resLower Service class" = "Cov: Intercept × Lower Service",
  "Cov: country2 (Intercept) dclass6resRoutine nonmanual class" = "Cov: Intercept × Routine Nonmanual",
  "Cov: country2 (Intercept) dclass6resSelf-employed" = "Cov: Intercept × Self-employed",
  "Cov: country2 (Intercept) dclass6resSkilled working class" = "Cov: Intercept × Skilled Working Class",
  "Cov: country2 (Intercept) dclass6resUnskilled working class" = "Cov: Intercept × Unskilled Working Class",

  # Covariances with Homogeneity
  "Cov: country2 homclass3_V_gc dclass6resLower Service class" = "Cov: Homogeneity × Lower Service",
  "Cov: country2 homclass3_V_gc dclass6resRoutine nonmanual class" = "Cov: Homogeneity × Routine Nonmanual",
  "Cov: country2 homclass3_V_gc dclass6resSelf-employed" = "Cov: Homogeneity × Self-employed",
  "Cov: country2 homclass3_V_gc dclass6resSkilled working class" = "Cov: Homogeneity × Skilled Working Class",
  "Cov: country2 homclass3_V_gc dclass6resUnskilled working class" = "Cov: Homogeneity × Unskilled Working Class",

  # Covariances between classes
  "Cov: country2 dclass6resLower Service class dclass6resRoutine nonmanual class" = "Cov: Lower Service × Routine Nonmanual",
  "Cov: country2 dclass6resLower Service class dclass6resSelf-employed" = "Cov: Lower Service × Self-employed",
  "Cov: country2 dclass6resLower Service class dclass6resSkilled working class" = "Cov: Lower Service × Skilled Working Class",
  "Cov: country2 dclass6resLower Service class dclass6resUnskilled working class" = "Cov: Lower Service × Unskilled Working Class",
  "Cov: country2 dclass6resRoutine nonmanual class dclass6resSelf-employed" = "Cov: Routine Nonmanual × Self-employed",
  "Cov: country2 dclass6resRoutine nonmanual class dclass6resSkilled working class" = "Cov: Routine Nonmanual × Skilled Working Class",
  "Cov: country2 dclass6resRoutine nonmanual class dclass6resUnskilled working class" = "Cov: Routine Nonmanual × Unskilled Working Class",
  "Cov: country2 dclass6resSelf-employed dclass6resSkilled working class" = "Cov: Self-employed × Skilled Working Class",
  "Cov: country2 dclass6resSelf-employed dclass6resUnskilled working class" = "Cov: Self-employed × Unskilled Working Class",
  "Cov: country2 dclass6resSkilled working class dclass6resUnskilled working class" = "Cov: Skilled Working Class × Unskilled Working Class",
  "Var: Residual" = "Residual variance"
)
      )
```

```{r tableS6}
dfreg$d10d1 <- as.numeric(scale(dfreg$d10d1))
dfreg$wid_p90p50 <- as.numeric(scale(dfreg$wid_p90p50))
int_xlvl_r9010 <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         homclass3_V_gc*
         dclass3res_V*
         d10d1 + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)

int_xlvl_wid_p90p50 <- 
  lmer(egal~ 
         homclass3_V_gc+
         know_total_gc+
         dclass3res_V+
         female_gc+agenum_gc+
         homclass3_V_gc*
         dclass3res_V*
         wid_p90p50 + 
         (homclass3_V_gc+dclass3res_V|country2),data=dfreg,weights = WEIGHT)




screenreg(list(int_xlvl_r9010,int_xlvl_wid_p90p50))

ccoef <- list(
  "dclass3res_VIntermediate class (III+IV)" = "Intermediate Class",
  "dclass3res_VWorking Class (V+VI+VII)" = "Working Class",
  
  homclass3_V_gc = "Class-Based Network Homogeneity (CWC)",
  
  "homclass3_V_gc:dclass3res_VIntermediate class (III+IV)" = "Intermediate Class × Homogeneity",
  "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII)" = "Working Class × Homogeneity",
  
  d10d1 = "Inequality (Ratio 90/10)",
  "homclass3_V_gc:d10d1" = "Homogeneity × Inequality (R90/10)",
  "dclass3res_VIntermediate class (III+IV):d10d1" = "Intermediate Class × Inequality (R90/10)",
  "dclass3res_VWorking Class (V+VI+VII):d10d1" = "Working Class × Inequality (R90/10)",  
  "homclass3_V_gc:dclass3res_VIntermediate class (III+IV):d10d1" = "Intermediate Class × Homogeneity × Inequality (R90/10)",
  "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII):d10d1" = "Working Class × Homogeneity × Inequality (R90/10)",
  
  wid_p90p50 = "Inequality (Ratio 90/50)",
  "homclass3_V_gc:wid_p90p50" = "Homogeneity × Inequality (R90/50)",
  "dclass3res_VIntermediate class (III+IV):wid_p90p50" = "Intermediate Class × Inequality (R90/50)",
  "dclass3res_VWorking Class (V+VI+VII):wid_p90p50" = "Working Class × Inequality (R90/50)",   
  "homclass3_V_gc:dclass3res_VIntermediate class (III+IV):wid_p90p50" = "Intermediate Class × Homogeneity × Inequality (R90/50)",
  "homclass3_V_gc:dclass3res_VWorking Class (V+VI+VII):wid_p90p50" = "Working Class × Homogeneity × Inequality (R90/50)"
)

htmlreg(list(int_xlvl_r9010,int_xlvl_wid_p90p50),stars = c(0.001, 0.01, 0.05, 0.1),scalebox = 0.7,symbol = ".",
       file = here::here("output/tables/tableS06.html"),
       caption = paste("Table S6: Robustness analysis of the multilevel models for income inequality ratios, network homogeneity and redistributive preferences"),
       custom.note = " Note: All individual level controls centered within cluster (group mean).  Standard errors in parentheses. %stars",
       caption.above = T,leading.zero = T,single.row = T,threeparttable = T,
       custom.coef.map = ccoef,
      # groups = list("Social Class (Ref.= Service Class)" = 2:3,"Homogeneity x Social Class"=7:8),
      custom.gof.rows = list("Controls"=rep("Yes",2)),
      include.loglik = FALSE,include.aic = FALSE,
      custom.gof.names = c(NA,NA,"Num. groups",
                           "Var: Country (Intercept)",
                           "Var: Country Homogeneity",
                           "Var: Country Intermediate Class ",
                           "Var: Country Working Class",
                           "Cov: Country (Intercept), Homogeneity",
                           "Cov: Country (Intercept), Intermediate Class ",
                           "Cov: Country (Intercept), Working Class",
                           "Cov: Country Homogeneity, Intermediate Class",
                           "Cov: Country Homogeneity, Working Class",
                           "Cov: Country Intermediate Class, Working Class",
                           "Var: Residual")
      )

```

```{r tableS7}
tab_countries<- 
  dfreg %>% 
  group_by(country3) %>% 
  summarise(gini_disp=mean(gini_disp)) %>% 
  mutate(gini_q05=ntile(gini_disp,5),
         gini_d10=ntile(gini_disp,10)) %>% as.data.frame() %>% 
  arrange(gini_d10) %>% 
  dplyr::select(-gini_disp)

dfreg <- dfreg %>% left_join(tab_countries,by="country3")


## Model -  EGP3 - DIGCLASS - V to Working -----------------------------$

library(estimatr)

# Quintile 1 (low)
digclas1_VW_Q1<- lm(egal~1 + dclass3res_V*homclass3_V_res +know_total+female+agenum+factor(country2),data=subset(dfreg,gini_q05==1),weights = WEIGHT)

# Quintile 2 (mid-low)
digclas1_VW_Q2<- lm(egal~1 + dclass3res_V*homclass3_V_res +know_total+female+agenum+ factor(country2),data=subset(dfreg,gini_q05==2),weights = WEIGHT)

# Quintile 3 (middle)
digclas1_VW_Q3<- lm(egal~1 + dclass3res_V*homclass3_V_res +know_total+female+agenum+ factor(country2),data=subset(dfreg,gini_q05==3),weights = WEIGHT)

# Quintile 4 (middle-high)
digclas1_VW_Q4<- lm(egal~1 + dclass3res_V*homclass3_V_res +know_total+female+agenum +factor(country2),data=subset(dfreg,gini_q05==4),weights = WEIGHT)

# Quintile 5 (high)
digclas1_VW_Q5<- lm(egal~1 + dclass3res_V*homclass3_V_res +know_total+female+agenum+ factor(country2),data=subset(dfreg,gini_q05==5),weights = WEIGHT)

digclas1_Q05 <- list("Q1"=digclas1_VW_Q1,"Q2"=digclas1_VW_Q2,"Q3"=digclas1_VW_Q3,"Q4"=digclas1_VW_Q4,"Q5"=digclas1_VW_Q5)

ccoef <- list("(Intercept)" = "(Intercept)",
              "dclass3res_VIntermediate class (III+IV)"="Intermediate Class",
              "dclass3res_VWorking Class (V+VI+VII)"="Working Class",
              homclass3_V_res="Class-based network homogeneity",
              "dclass3res_VIntermediate class (III+IV):homclass3_V_res" = "Intermediate Class × Homogeneity",
              "dclass3res_VWorking Class (V+VI+VII):homclass3_V_res" = "Working Class × Homogeneity")
# header <- list("Q1"=1:2,"Q2"=3:4,"Q3"=5:6,"Q4"=7:8,"Q5"=9:10)
htmlreg(digclas1_Q05,stars = c(0.001, 0.01, 0.05, 0.1),single.row = T,include.ci = FALSE,
        file = here::here("output/tables/tableS07.html"),
        caption = "Table S7: Fixed effects linear regression models for Class-based network segregation and Redistributive Preferences by Income Inequality Quintiles",
        custom.note = "Note: Standard errors in parentheses. %stars",
        caption.above = T,leading.zero = T,
        custom.coef.map = ccoef,
      groups = list("Social Class (Ref.= Service Class)" = 2:3,
                    "Social Class × Homogeneity"=5:6),
      custom.gof.rows = list("Country FE"=rep("Yes",5)),
      # include.loglik = FALSE,include.aic = FALSE,
      # custom.gof.names = c(NA,NA,"Num. groups","Var: Country (Intercept)","Var: Residual")
      )
```
